==> 08-WillowHt.Rmd <==
---
title: "Willow Height - Data Processing Report"
output:
  html_document: 
    fig_caption: yes
    fig_height: 8
    number_sections: yes
    theme: journal
    toc: yes
    toc_depth: 2
    toc_float: yes
  word_document: default
---

```{r setup, include=FALSE}

library(knitr)
opts_chunk$set(echo = TRUE, warning = TRUE, message = FALSE, fig.align = 'center')
# opts_knit$set(root.dir=normalizePath('../')) # required if Rmd is nested below the project directory

```

**Updated:** `r format(Sys.time(), '%d %B, %Y')`

# Introduction

This document evaluates willow height measurement data from project inception through 2019. Data include those archived on the "Digital Collections of Colorado" (DCC) library site and more recent (2015-2019) data.

DCC links:

Plant level fall current annual growth and height on experimental plots in Yellowstone's Northern Range, 2002 - 2015: <http://hdl.handle.net/10217/173649>

Plant level fall current annual growth and height on observation plots in Yellowstone's Northern Range, 2008 - 2015: <http://hdl.handle.net/10217/173657>

```{r,echo=FALSE}
suppressPackageStartupMessages(library(tidyverse))
library(fs)
suppressPackageStartupMessages(library(sf))
# library(raster)
suppressPackageStartupMessages(library(janitor))
suppressPackageStartupMessages(library(readxl))
# library(glue)
suppressPackageStartupMessages(library(mapview))
# library(ggmap)
# library(ggrepel)
suppressPackageStartupMessages(library(viridis))
library(ggExtra)
library(ggrepel)
library(DT)
# library(kableExtra)
# suppressPackageStartupMessages(library(compare))
suppressPackageStartupMessages(library(skimr)) ## some useful functions
suppressPackageStartupMessages(library(dataMaid))
library(lubridate)
library(gt)
library(summarytools)
```

# Setup

## Functions

```{r, functs}
## data import
read_sheets <- function(file){
  xlsx_file <- file
  xlsx_file %>%
    excel_sheets() %>%
    set_names() %>%
    map_df(read_excel, path = xlsx_file, .id = 'sheet_name', trim_ws = TRUE, skip = 1, col_types = "text", range = "A2:DW700") %>% 
    mutate(file_name = file) %>% 
    select(file_name, sheet_name, everything())
}

## read xlsx for production files
read_sheets_prod <- function(file){
  xlsx_file <- file
  xlsx_file %>%
    excel_sheets() %>%
    set_names() %>%
    map_df(read_excel, path = xlsx_file, .id = 'sheet_name', trim_ws = TRUE, skip = 1, col_types = "text", range = "A2:DW700") %>% 
    mutate(file_name = file) %>% 
    select(file_name, sheet_name, everything())
}

read_sheets_prod18 <- function(file){
  xlsx_file <- file
  xlsx_file %>%
    excel_sheets() %>%
    set_names() %>%
    # map_df(read_excel, path = xlsx_file, .id = 'sheet_name', trim_ws = TRUE, skip = 1, col_types = "text", range = "A2:Dw136") %>%
    map_df(read_excel, path = xlsx_file, .id = 'sheet_name', trim_ws = TRUE, skip = 1, col_types = "text", range = "A2:DW700") %>%
    mutate(file_name = file) %>% 
    select(file_name, sheet_name, everything())
}


read_sheets_util <- function(file){
  xlsx_file <- file
  xlsx_file %>%
    excel_sheets() %>%
    set_names() %>%
    map_df(read_excel, path = xlsx_file, .id = 'sheet_name', trim_ws = TRUE, skip = 1, col_types = "text", range = "A2:DW700") %>% 
    mutate(file_name = file) %>% 
    select(file_name, sheet_name, everything())
}

## read xlsx then csv cache
read_then_csv <- function(sheet, path) {
  pathbase <- path %>%
    basename() %>%
    tools::file_path_sans_ext()
  path %>%
    read_excel(sheet = sheet) %>% 
    write_csv(paste0(pathbase, "-", sheet, ".csv"))
}

## plotting
ggTile_yr_season_site2 <- function(df){
  df %>%
  group_by(yr, season, site2) %>%
  summarise(n= n()) %>%
  ggplot(aes(yr, site2)) +
  geom_tile(aes(fill = n), color = 'white') +
  viridis::scale_fill_viridis(option = "B") +
  facet_wrap(~season) +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 7)) +
  theme(axis.text.y = element_text(size = 7)) +
  facet_wrap(~season)}

## tile - source
ggTile_source_yr_season_site2 <- function(df){
  df %>%
  group_by(source, yr, season, site2) %>%
  tally() %>%
  ggplot(aes(yr, site2)) +
  geom_tile(aes(fill = n), color = 'white') +
  viridis::scale_fill_viridis(option = "B") +
  facet_grid(source ~ season) +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 7)) +
  theme(axis.text.y = element_text(size = 7))
}


ggTile_yr_season_site <- function(df){
  df %>%
  group_by(yr, season, site) %>%
  summarise(n= n()) %>%
  ggplot(aes(yr, site)) +
  geom_tile(aes(fill = n), color = 'white') +
  viridis::scale_fill_viridis(option = "D") +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 7)) +
  theme(axis.text.y = element_text(size = 7)) +
  facet_wrap(~season)} 

##
ggTile_yr_season_plot <- function(df){
  df %>%
  group_by(yr, plot, season, site) %>%
  summarise(n= n()) %>%
  ggplot(aes(plot, site)) +
  geom_tile(aes(fill = n), color = 'white') +
  viridis::scale_fill_viridis(option = "B") +
  facet_wrap(~season) +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 9)) +
  theme(axis.text.y = element_text(size = 8)) +
  facet_wrap(yr~season)
  } 
#
##
ggTile_yr_season_plot_alt1 <- function(df){
  df %>%
  group_by(yr, plot, season, site) %>%
  summarise(n= n()) %>%
  ggplot(aes(yr, site)) +
  geom_tile(aes(fill = n), color = 'white') +
  viridis::scale_fill_viridis(option = "A") +
  # facet_wrap(~season) +
  facet_grid(season~plot) +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 9)) +
  theme(axis.text.y = element_text(size = 8))
} 

### ggpoint
ggPoint_yr_site_season_plot <- function(df){
  df %>%
  ggplot(aes(yr, pl_ht_cm)) +
    geom_point(aes(color = plot, shape = plot)) +
    facet_wrap(~site, ncol = 4) +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, size = 9)) +
  theme(axis.text.y = element_text(size = 8)) +
    viridis::scale_color_viridis(discrete = TRUE, option = "C") +
    labs(y = "Height (cm)")
} 


### gt table summary
tbqc_exp <- function(df){
  df %>%
    filter(plot != "obs") %>% 
    group_by(site,yr, season, plot) %>% 
    summarise(n = count(), ht_mean = mean(pl_ht_cm, na.rm = TRUE), ht_sd = sd(pl_ht_cm, na.rm = TRUE)) %>% 
    gt()
}

tbqc_obs <- function(df){
  df %>%
    filter(plot == "obs") %>% 
    group_by(site,yr, season, plot) %>% 
    summarise(n = count(), ht_mean = mean(pl_ht_cm, na.rm = TRUE), ht_sd = sd(pl_ht_cm, na.rm = TRUE)) %>% 
    gt()
}

### Table funcs
gt_tally_site_plot_yr_season <- function(df){
  df %>%
  group_by(yr, plot, season, site) %>%
  tally() %>%
    gt::gt() %>% 
    tab_header(title = "tally")
} 

```

```{r}
### Table functions
## gt field names
names.gt <- function(df){
  df %>% 
  names() %>% enframe %>% gt::gt() %>% 
    tab_header(title = "field names")
}

names.dt <- function(df){
  df %>% 
  names() %>% enframe %>% 
    DT::datatable(caption = "field names")
}


```

## Lookup tables

```{r}

## wilid_full spp lookup
lu_wilid_full_spp <- read_csv("./data/lu_wilid_full_spp.csv")  
## bv occ
lu_site_beav <- read_csv("./data/lu_site_site2_beav.csv")


### try to add spp info to the 18 and 19 
wilid.spp.lu2 <- read_csv("./data/wilid_spp_lu_v2.csv") %>% 
  select(c(species, wilid_full)) %>% 
  distinct()

## change ref system to start with lu
lu.wilid.spp.v2 <- read_csv("./data/wilid_spp_lu_v2.csv") %>% 
  select(c(species, wilid_full)) %>% 
  distinct()

# lu.wilid.spp.v2 %>% 
#   tabyl(wilid_full)

plot.lu <- read_csv("./data/plot_lookup_master.csv",col_types = "ccccc")

## read in lu
lu_spp_site2 <- read_csv("./data/lu_spp_site2.csv")

```

# Data import and wrangling

## Data files from Tom

TH provided a separate file with data 2001-2015 data. Unclear how this was developed or how it relates to either the DCC or records gathered from the various prod and util files.

```{r}

th.01.15 <- read_csv("./data/Yell_data_from_Hobbs_drive/Master data file/WillowHt_2001_2015.csv")

th.01.15 <- th.01.15 %>%
  mutate(plot = case_when(dam == 0 & browse == 0 ~ "cx",
                           dam == 1 & browse == 0 ~ "dx",
                           dam == 0 & browse == 1 ~ "cc",
                           dam == 1 & browse == 1 ~ "dc",
                           TRUE ~ "obs"
                           )) %>% 
  select(-c(ID,exp,dam,browse)) %>% 
  rename(yr = year) %>% 
  mutate(yr = as.character(yr)) %>% 
  rename(wilid = willid) %>% 
  rename(pl_ht_cm = plantht)

## add "source" field to allow tracking
th.01.15 <- th.01.15 %>%
  mutate(source = "WillowHt_2001_2015.csv")

```

```{r}
th.01.15 %>% 
  ggTile_yr_season_site() +
  labs(title = "2001-2015 data from TH file", caption = "WillowHt_2001_2015.csv")

```

No spring data here, except for 2001.

## DCC

```{r, message=FALSE, warning=FALSE}

## revised 20180228
# w_ht <- read_excel("data/Select_files/willow_height/WillowHt_2001_2017_Final.xlsx")

# dataframe comparisons
# setdiff(bigFrame, smallFrame) # dplyr solution for comparison
# comp <- setdiff(w_ht,w_ht2)

# Using functions in "comparison" package
# comparison <- compare(w_ht,w_ht2,allowAll=TRUE)

## Continuing with the file downloaded from the GDr 20180321
w_ht <- read_excel("./data/raw/ht/2016_2017_Data_Entry_Format_WillowHt_2001_2015.xlsx")

## add "source" for tracking
w_ht <- w_ht %>% 
  mutate(source = "2016_2017_Data_Entry_Format_WillowHt_2001_2015.xlsx")

w_ht %>%
  # distinct(season)
  mutate(year = as.character(year)) %>% 
  group_by(site, season, year) %>% 
  tally() %>% 
  ggplot(aes(year, site)) +
  geom_tile(aes(fill = n)) +
  facet_wrap(~season) +
  labs(caption = "data/raw/ht/2016_2017_Data_Entry_Format_WillowHt_2001_2015.xlsx") +
  theme(axis.text.x = element_text(angle = 55, hjust = 1))

## yet another file possibly with data...
w_ht01_17 <- read_excel("data/raw/ht/WillowHt_2001_2017_QAQCd_Version.xlsx")

w_ht01_17 <- w_ht01_17 %>% 
  mutate(source = "WillowHt_2001_2017_QAQCd_Version.xlsx")

w_ht <- bind_rows(w_ht, w_ht01_17) %>% 
  # select(-c(ID, exp, dam, browse)) %>% 
  filter(plantht != "NA") %>%
  rename(pl_ht_cm = plantht) %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>%   distinct()

## one of dozens of differences between files: willid vs wilid. For consistency with latter files, I will go with wilid
w_ht <- w_ht %>% 
  rename(wilid = willid)
### these files have no species id
```


```{r}
### w_ht issue

w_ht <- w_ht %>%
  filter(pl_ht_cm != "na" & pl_ht_cm != "missing" & pl_ht_cm != "NA") %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>% 
  mutate(wilid_full = paste0(wilid,"=", site)) 

## don't think I need this bc use of lu
w_ht <- w_ht %>%
#   # rename(pl_ht_cm = plantht) %>% 
  mutate(plot = case_when(dam == 1 & browse == 0 ~ "dx",
                           dam == 1 & browse == 1 ~ "dc",
                           dam == 0 & browse == 0 ~ "cx",
                           dam == 0 & browse == 1 & exp == 1 ~ "cc",
                           dam == 0 & browse == 1 & exp != 1 ~ "obs",
                           dam == 0 & browse == 1 ~ "cc",
                           dam == 2 & browse == 1 ~ "obs",
                           TRUE ~ "oth")) %>% 
  mutate(site2 = paste0(site, plot, sep = '-'))
  

# not sure on the interpretation of the "2"
# w_ht %>% distinct(dam) # Three values: 0,1,2. Expected only 0,1.
# added the site2 field to match that used in well data 

# Tip: if you are testing a scalar, use if(). Testing a vector against a single condition, dplyr::if_else. Testing a vector against multiple conditions, use case_when.
  
```

```{r}

w_ht <- w_ht %>% 
  mutate_if(.predicate = is.character,.funs=tolower)

w_ht <- w_ht %>% 
  mutate(site = case_when(site == "halibuff"~"hailbuf",
                          TRUE ~ site))

w_ht %>% 
  visdat::vis_dat()

```


> Interogating other raw files looking for spring data

```{r, eval=FALSE}
w_ht01_17 %>% 
  group_by(year, season, site) %>%
  tally() %>% 
  ggplot(aes(as.character(year), season)) +
  geom_tile(aes(fill = n)) +
  labs(caption = "WillowHt_2001_2017_QAQCd_Version.xlsx")
# > Still missing spring data...
```

### Raw DCC spring

```{r}
## MORE SPRING DATA?!
## the experimental
nsf.ht.spr <- read_delim("./data/NSF_DataArchive20180208/Plant_level_spring_current_annual_growth_and_height_on_experimental_plot/willows-plantCAGspring2003-2015.txt", delim = " ")

## add "source" for tracking
nsf.ht.spr <- nsf.ht.spr %>% 
  mutate(source = "DCC - willows-plantCAGspring2003-2015.txt")

###
nsf.ht.spr.obs <- read_delim("./data/NSF_DataArchive20180208/Plant_level_spring_current_annual_growth_and_height_on_observation_plot/OBS-plantCAGspring2009-2013.txt", delim = " ", col_types = "cccccddcccc")

## add "source" for tracking
nsf.ht.spr.obs <- nsf.ht.spr.obs %>% 
  mutate(source = "DCC - OBS-plantCAGspring2009-2013.txt")

# clean. rename. These are (surprise, surprise) inconsistently named
nsf.ht.spr <- nsf.ht.spr %>% 
  select(year, site, plot, species, willid, plantht, source) %>% 
  rename(yr = year, wilid = willid, pl_ht_cm = plantht) %>% 
  mutate(season = "spring")

nsf.ht.spr.obs <- nsf.ht.spr.obs %>% 
  select(year, site, plot, species, willid, plantht, source) %>% 
  rename(yr = year, wilid = willid, pl_ht_cm = plantht)

nsf.ht.spr <- nsf.ht.spr %>% 
  mutate(season = "spring") %>% 
  mutate(yr = as.character(yr)) %>% 
  mutate(wilid = as.character(wilid))

nsf.ht.spr.obs <- nsf.ht.spr.obs %>% 
  filter(!is.na(pl_ht_cm)) %>% 
  mutate(season = "spring") %>% 
  mutate(plot = "obs")

## combine obs and exp for spring
nsf.ht.spr <- bind_rows(nsf.ht.spr,nsf.ht.spr.obs)

# clean
nsf.ht.spr <- nsf.ht.spr %>%
  mutate(plot = if_else(is.na(site),"obs",plot)) %>%
  filter(!is.na(plot))

```

Plots for QA on DCC spring files.

```{r}
nsf.ht.spr %>% 
  # distinct(site)
  visdat::vis_miss() +
  labs(title = "QA plot on raw DCC data", caption ="OBS-plantCAGspring2009-2013.txt, willows-plantCAGspring2003-2015.txt")

nsf.ht.spr %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  ggTile_yr_season_site2() +
  labs(title = "QA plot on raw DCC data", caption ="OBS-plantCAGspring2009-2013.txt, willows-plantCAGspring2003-2015.txt")

```
### Raw DCC fall

```{r}
## Fa raw DCC
nsf.ht.fa.exp <- read_delim("./data/NSF_DataArchive20180208/Plant_level_fall_current_annual_growth_and_height_on_experimental_plot/willows-plantCAGfall2002-2015.txt", delim = " ") %>% 
  select(year, site, plot, species, willid, height) %>% 
  rename(yr = year, wilid = willid, pl_ht_cm = height) %>% 
  mutate(season = "fall")

## add "source" for tracking
nsf.ht.fa.exp <- nsf.ht.fa.exp %>% 
  mutate(source = "DCC - willows-plantCAGfall2002-2015.txt")

#### obs DCC
nsf.ht.fa.obs <- read_delim("./data/NSF_DataArchive20180208/Plant_level_fall_current_annual_growth_and_height_on_observation_plot/OBS_shootCAGfall08-15.txt", delim = " ") %>% 
  select(year, site, plot, species, willid, height) %>% 
  rename(yr = year, wilid = willid, pl_ht_cm = height) %>% 
  mutate(season = "fall")

## add "source" for tracking
nsf.ht.fa.obs <- nsf.ht.fa.obs %>% 
  mutate(source = "DCC - OBS_shootCAGfall08-15.txt")


```

Combine raw FALL DCC exp and obs
```{r}
## combine raw FALL dcc exp and obs
nsf.ht.fa.expobs <- bind_rows(nsf.ht.fa.exp, nsf.ht.fa.obs) %>% 
  distinct() %>% 
  mutate(yr = as.character(yr), wilid = as.character(wilid))

## dcc combined
dcc.ht.fa.spr <- bind_rows(nsf.ht.fa.expobs, nsf.ht.spr) %>%
  mutate(wilid = as.character(wilid)) %>% 
  distinct()

# qa plot
# visdat::vis_miss(dcc.ht.fa.spr)
```

```{r}
# clean. remove na for pl_ht_cm
dcc.ht.fa.spr <- dcc.ht.fa.spr %>% 
  filter(!is.na(pl_ht_cm))

## plot is NA if not exp. Need it explicit if to be used to constitute a wilid_full attribute through paste0
dcc.ht.fa.spr <- dcc.ht.fa.spr %>% 
  mutate(plot = if_else(is.na(plot),"obs",plot)) %>% 
  mutate(wilid_full = paste0(site,"-",plot, "-", wilid)) 

## add 'site2'
dcc.ht.fa.spr <- dcc.ht.fa.spr %>% 
  mutate(site2 = paste0(site,"-",plot))

```

```{r}
## export
# dcc.wilid.spp.lu %>% 
#   write_csv("./data/dcc_wilid_spp_lu.csv")
```


QA plot - combined fall and spring DCC data 
```{r}
dcc.ht.fa.spr %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  ggTile_yr_season_site2() +
  labs(title = "Raw combined spring and fall DCC", caption = "dcc.ht.fa.spr")
```

QA plot - combined fall and spring DCC data by source file
```{r, fig.height=10, fig.width=6}
## qa plot
ggTile_source_yr_season_site2(dcc.ht.fa.spr)+
  labs(title = "Raw combined spring and fall DCC", caption = "dcc.ht.fa.spr")
```


### Ht from sprcag\*.csv files

```{r}
## see if there is data in the sprcag*.csv files
# updated as of 20191203
files_util_01_15 <- fs::dir_ls("./data/Yell_data_from_Hobbs_drive/Utilization all years", recurse = FALSE, glob = "*.csv")

sprcag.files <- enframe(files_util_01_15) %>%
  rename(path = name) %>% 
  select(-value) %>%
  separate(data = ., col = path, into = c("f1","f2","f3","f4","source"), sep = "/", remove = FALSE) %>% 
  select(-starts_with("f"))

sprcag.files %>%
  gt() %>% 
  tab_header(title = "sprcag*.csv files")

sprcag.files <- files_util_01_15 %>%
  map_df(~read_csv(., col_types = cols(.default = "c")))

sprcag.files.ht <- sprcag.files %>%
  mutate(season = "spring") %>% 
  rename(wilid = willid) %>% 
  rename(yr = year) %>% 
  rename(pl_ht_cm = plantht) %>% 
  select(yr,season, site,plot,wilid,pl_ht_cm) %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>%
  filter(!is.na(pl_ht_cm)) %>% 
  distinct() %>% 
  mutate(source = "sprcag*.csv")

```

```{r, eval=FALSE}
## qa plot
sprcag.files.ht %>% 
  visdat::vis_dat()
```


```{r}
sprcag.files.ht %>% 
  mutate(yr = as.integer(yr)) %>% 
  ggTile_yr_season_site() +
  labs(title = "Combined 'sprcag*.csv' files")

```

> Note missing data

> Observation sites?

### Ht from prod\*.csv files

```{r}
files_prod_01_15 <- fs::dir_ls("./data/Yell_data_from_Hobbs_drive/Production all years", recurse = FALSE, glob = "*.csv")
# regex
# files_util_01_15 <- fs::dir_ls("./data/Yell_data_from_Hobbs_drive/Utilization all years", regexp = "\\.csv$")

prod.files <- files_prod_01_15 %>%
  map_df(~read_csv(., col_types = cols(.default = "c")))

# prod.files %>% 
#   glimpse()

### clean and create a new obj
prod.files.ht <- prod.files %>%
  mutate(season = "fall") %>% 
  rename(wilid = willid) %>% 
  rename(yr = year) %>% 
  rename(pl_ht_cm = height) %>% 
  select(yr,season, site,plot,wilid,pl_ht_cm) %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>%
  filter(!is.na(pl_ht_cm)) %>% 
  distinct() %>% 
  mutate(source = "prod*.csv")

```

```{r}
enframe(files_prod_01_15) %>%
  rename(path = name) %>% 
  select(-value) %>%
  separate(data = ., col = path, into = c("f1","f2","f3","f4","prod_file"), sep = "/", remove = FALSE) %>% 
  select(-starts_with("f")) %>% 
  gt()

```

```{r}
## combine production and sprcag obj
prod.sprcag.ht <- bind_rows(prod.files.ht,sprcag.files.ht)
```

```{r}
prod.sprcag.ht %>% 
  mutate(site2 = paste0(site,"-", plot)) %>% 
  ggTile_yr_season_site2() +
  labs(subtitle = "Spring and Fall Data from sprcag*.csv and prod*.csv files", caption = "prod.sprcag.ht")

prod.sprcag.ht %>% 
  mutate(site2 = paste0(site,"-", plot)) %>% 
  ggTile_yr_season_site()

```

```{r}
# wilids from sprcag and prod files
sprcag.files.ht %>% 
  distinct(season,site,plot,wilid) %>% 
  mutate(wilid = as.integer(wilid)) %>% 
  arrange(wilid) %>% 
  datatable(rownames = FALSE, filter = 'bottom')

```

```{r}
prod.files.ht %>% 
  distinct(season,site,plot,wilid) %>% 
  mutate(wilid = as.integer(wilid)) %>% 
  arrange(wilid) %>% 
  datatable(rownames = FALSE, filter = 'bottom')

# > Note: where are the obs sites?
# 
# > Error in files with eb4-cc?
```

## 2016 Production

```{r}
## read in raw file
p16.raw <- read_excel("./data/raw/production/2016_Raw_Production_20180421.xlsx", col_types = "text") %>% 
  janitor::clean_names()

## rename inconsistently named columns
p16.raw <- p16.raw %>% 
  rename(wilid = wildid) %>% 
  rename(species = spp) %>% 
  rename(pl_ht_cm = height_cm) %>%
  mutate(season = "fall") %>% 
  mutate(yr = as.character("2016"))

## trim white space
p16.raw <- p16.raw %>% 
  mutate(wilid = str_trim(string = .$wilid))

## plot is NA for obs sites. Making explicit
p16.raw <- p16.raw %>% 
  mutate(plot = if_else(is.na(plot),"obs",plot)) %>%
  mutate(site2 = paste0(site, "-", "plot"))

# coercing wilid to integer. This induces NA because of non-numeric entries (e.g., wilid = "?"). I assume these are to be culled
p16.raw <- p16.raw %>% 
  mutate(wilid = as.integer(wilid))

## create wilid full
p16.raw <- p16.raw %>% 
  mutate(wilid_full = paste0(site,"-",plot,"-",wilid))

```

```{r}
# p16.raw %>% 
#   distinct(site) %>%
#   View()

# p16.raw %>% 
#   distinct(site, wilid) %>%
#   gt() %>% 
#   tab_header(title = "distinct site/wilid", subtitle = "2016_Raw_Production_20180421.xlsx")
# 
# p16.raw %>% 
#   distinct(site, wilid_full) %>%
#   gt() %>% 
#   tab_header(title = "distinct site/wilid", subtitle = "2016_Raw_Production_20180421.xlsx")

p16.raw %>% 
  distinct(site, wilid_full) %>%
  datatable(rownames = FALSE, caption = "2016_Raw_Production_20180421.xlsx")

```

```{r}
p16.raw %>% 
  distinct(site, wilid) %>% 
  arrange(site) %>% 
  datatable(rownames = FALSE, caption = "Distinct wilid in raw 2016 production excel file:2016_Raw_Production_20180421.xlsx",filter = 'top')

p16.raw %>% 
  group_by(wilid) %>% 
  summarise(n.distinct.sites = n_distinct(site))%>% 
  datatable(rownames = FALSE, caption = "Count of sites for each wilid value in raw 2016 production excel file: 2016_Raw_Production_20180421.xlsx")


```

```{r}

```

```{r, eval=TRUE}
# !!!! issue with wilid parsing!
# p16.raw <- p16.raw %>% 
#   mutate(wilid_full = str_trim(string = .$wilid_full)) %>%
#   # mutate(wilid.test = str_remove(wilid, ".0")) %>% 
#   mutate(wilid_full = str_sub(string = wilid_full,start =  1,-3))

```

```{r}
### Break out height from shoot level fields (e.g., sht_len_1)

p16.ht <- p16.raw %>%  
  dplyr::select(c(yr, wilid_full, site2, wilid, site, plot, species, pl_ht_cm,stem_id, live, season, notes))


# !!!! issue with wilid parsing!
p16.ht <- p16.ht %>% 
  mutate(wilid_full = str_trim(string = .$wilid_full)) %>%
  # mutate(wilid.test = str_remove(wilid, ".0")) %>% 
  mutate(wilid_full = str_sub(string = wilid_full,start =  1,-3))


# p16.ht <- p16.ht %>% 
#   left_join(., lu.wilid.spp.v2, by = "wilid_full") %>%
#   mutate(species = case_when(is.na(species.x)~species.y,
#                              is.na(species.y)~species.x,
#                              TRUE ~ "NA")) %>% 
#   select(-c(species.x, species.y))

# p16.ht <- 
#   p16.ht %>% 
#   # filter(!is.na(pl_ht_cm)) %>% 
#   filter(!is.na(wilid))

p16.ht %>% 
  visdat::vis_dat()

# u16.ht %>% 
#   filter(is.na(species))
```

```{r}
p16.ht <- p16.ht %>%
  # select(-c(species)) %>%
  mutate(yr = as.character("2016")) %>%
  mutate(season = "fall") %>%
  distinct()
  
p16.ht %>%   
  ggTile_yr_season_site() +
  labs(caption = "p16.ht")

```

```{r}
p16.ht %>% 
  group_by(wilid_full, site, plot) %>% 
  tally() %>% 
  datatable()

```

```{r}

p16.ht %>% 
  tabyl(site)

```

## 2016 Utilization

The following show what I know now: **NO SPRING DATA FOR 2016**

```{r}
# 2016 utilization
u16.raw <- read_excel("./data/raw/utilization/Raw_2016_Utilization/2016_Raw_Utilization.xlsx", col_types = "text") %>% 
  janitor::clean_names()

u16.raw <- u16.raw %>%
  rename(wilid = wildid) %>%
  mutate(wilid = str_trim(string = .$wilid)) %>%
  # mutate(wilid.test = str_remove(wilid, ".0")) %>% 
  mutate(wilid = as.numeric(wilid)) %>% 
  mutate(wilid = as.character(wilid)) %>%
  rename(species = spp) %>% 
  rename(pl_ht_cm = ht_cm) %>%
  rename(live_dead = live) %>% 
  mutate(season = "spring") %>% 
  mutate(yr = as.character(2016))

# some cleaning
u16.raw <- u16.raw %>% 
  mutate(site = if_else(site == "elk5 ?", "elk5",site)) %>%
  mutate(plot = if_else(is.na(plot),"obs",plot)) %>%
  mutate(site = if_else(site == "xtal","crystal",site)) %>% 
  mutate(site2 = paste0(site, "-", "plot")) 

u16.raw <- u16.raw %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm))

```

```{r}
u16.ht <- u16.raw %>% 
  select(c(yr, season, site, plot, wilid, pl_ht_cm)) %>% 
  distinct() %>% 
  filter(is.na(pl_ht_cm))

# u16.ht %>% 
#   visdat::vis_dat()

```

## 2017 Production

This reads in and munges data from "2017_Raw_Production_20180423.xlsx", accessed from the team drive after entry by LM+
```{r}

p17.raw <- read_excel("./data/raw/production/2017_Raw_Production_20180423.xlsx", col_types = "text") %>% 
  janitor::clean_names()

# p17.raw %>% glimpse()

p17.raw %>% distinct(plot)

## bad plot data entry
p17.raw <- p17.raw %>%
  # distinct(plot)
  mutate(plot = case_when(is.na(plot)~"obs", 
                          plot == "1" ~ "obs",
                          plot == "4" ~ "obs",
                          plot == "5" ~ "obs",
                          plot == "6" ~ "obs",
                          TRUE ~ plot)) 
p17.raw %>%
  tabyl(site, plot) %>% 
  gt() %>% 
  tab_header(title = "Count of observations in 2017_Raw_Production_20180423.xlsx")

```

```{r}
p17.raw <- p17.raw %>% 
  rename(wilid = wildid) %>% 
  rename(species = spp) %>% 
  rename(pl_ht_cm = height_cm) %>%
  rename(live_dead = live) %>% 
  mutate(season = "fall") %>% 
  mutate(yr = "2017") %>% 
  mutate(yr = as.character(yr)) %>% 
  mutate(source = "2017_Raw_Production_20180423.xlsx")

# some cleaning
p17.raw <- p17.raw %>% 
  # mutate(site = if_else(site == "elk5 ?", "elk5",site)) %>%
  mutate(plot = if_else(is.na(plot),"obs",plot)) %>%
  mutate(site = case_when(site == "wn" ~ "wb",
         TRUE ~ site)) %>% 
  mutate(site2 = paste0(site, "-", plot)) %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm))

```

```{r}
p17.ht <- p17.raw %>% 
  select(c(yr, season, site, site2, plot, species, wilid, pl_ht_cm, source)) %>% 
  mutate(wilid_full = paste0(site,"-",plot,"-",wilid))

```


```{r, eval=FALSE}
# quick check
p17.ht %>% 
  visdat::vis_dat()

p17.ht %>% 
  tabyl(wilid_full,plot) %>% 
  datatable()
```

```{r}
p17.ht %>% 
  distinct(site2, plot, wilid) %>% 
  datatable(caption = "2017_Raw_Production_20180423.xlsx")

p17.ht %>%
  group_by(wilid_full) %>%
  tally() %>% 
  datatable(caption = "Distinct wilid: 2017_Raw_Production_20180423.xlsx")

```

There are problematic wilid evident (e.g., wilid = ?, NA for wilid)

```{r, eval=FALSE}

## issue with obs plots having 
p17.ht %>% 
  ggTile_yr_season_site2() +
  labs(caption = "2017_Raw_Production_20180423.xlsx")

```

## 2017 Utilization
This reads in and munges data from *2017_Raw_Utilization.xlsx*, accessed from the team drive after entry by LM+

```{r}
# 2017 utilization
u17.raw <- read_excel("./data/raw/utilization/Raw_2017_Utilization/2017_Raw_Utilization.xlsx", col_types = "text") %>% 
  janitor::clean_names()

u17.raw <- u17.raw %>% 
  rename(wilid = wildid) %>% 
  rename(species = spp) %>% 
  rename(pl_ht_cm = ht_cm) %>%
  rename(live_dead = live) %>% 
  mutate(season = "spring") %>% 
  mutate(yr = as.character("2017"))

# some cleaning
u17.raw <- u17.raw %>% 
  mutate(site = if_else(site == "elk5 ?", "elk5", site)) %>%
  mutate(plot = if_else(is.na(plot),"obs",plot)) %>%
  mutate(site2 = paste0(site, "-", plot)) %>% 
  mutate(source = "2017_Raw_Utilization.xlsx")

u17.raw <- u17.raw %>%
  filter(!is.na(pl_ht_cm)) %>%  
  select(c(yr, site, site2, plot, wilid, season, yr, pl_ht_cm, source)) %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) # 

# seems like an obvious outlier, but will retain so TH can decide
# u17.raw %>% 
  # filter(pl_ht_cm < 800) # seems like an obvious outlier, but will retain so TH can decide

```

```{r}
u17.ht <- u17.raw %>%
  distinct()

u17.ht %>% 
  filter(is.na(wilid)) %>% 
  gt() %>% 
  tab_header(title = "Missing wilid from 2017_Raw_Utilization.xlsx")

```

```{r}
u17.ht <- u17.ht %>% 
  mutate(plot = case_when(site2 == "elk4-35" ~"obs",
                          TRUE ~ plot)) %>%
  mutate(plot = case_when(site2 == "elk3-35" ~"obs",
                          TRUE ~ plot)) %>%
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid))

```

QA plots:
```{r}
## qa plots
u17.ht %>% 
  ggTile_yr_season_site2() +
  labs(caption = "2017_Raw_Utilization.xlsx")

u17.ht %>% 
  visdat::vis_dat() +
  labs(caption = "2017_Raw_Utilization.xlsx")

```

```{r, eval = FALSE, fig.height=8}
# *u17.ht*
gg.u17 <- u17.ht %>% 
  ggplot(aes(x = site2, y = pl_ht_cm)) +
  geom_point(aes(color = plot))  +
  coord_flip()

plotly::ggplotly(gg.u17)
# this plot reveals a mis-entered ht of 3119. I checked the raw file it should be 319

```

```{r}
# plot reveals a mis-entered ht of 3119. I checked the raw file it should be 319. Change
u17.ht <- u17.ht %>% 
  mutate(pl_ht_cm = case_when(pl_ht_cm == 3119 ~ 319,
         TRUE ~ pl_ht_cm))

```

## 2018 Production

```{r}
# data updated as of 20191203
files2018 <- fs::dir_ls("./data/raw/production/Raw_2018_Production", recurse = FALSE, glob = "*.xlsx")

## 21 different files! -- some of which may have multi tabs
## split out the obs

obs.files2018 <- fs::dir_ls("./data/raw/production/Raw_2018_Production", recurse = FALSE, glob = "*obs_2018_Production.xlsx")

# 2018 - 4 of 21 files multi tabbed
exp.files2018 <- fs::dir_ls("./data/raw/production/Raw_2018_Production", recurse = FALSE, glob = "*exp_2018_Production.xlsx")

#### read in all 18 data
# p18 <- files2018 %>%
#   map_df(~ read_sheets(.))

#####
read_sheets_prod18exp <- function(file){
  xlsx_file <- file
  xlsx_file %>%
    excel_sheets() %>%
    set_names() %>% 
    map_df(read_excel, path = xlsx_file, .id = 'sheet_name', trim_ws = TRUE, skip = 1, col_types = "text", range = "A2:DW200") %>%
    mutate(file_name = file) %>%
    select(file_name, sheet_name, everything())
}
#####

p18 <- files2018 %>% 
  map_df(~ read_sheets_prod18(.)) %>% 
  clean_names() %>% 
  mutate(source = str_remove(file_name, "./data/raw/production/Raw_2018_Production/"))

## create a "source" field for tracking

p18.obs <- obs.files2018 %>% 
  map_df(~ read_sheets_prod18(.)) %>% 
  clean_names() %>% 
  mutate(source = str_remove(file_name, "./data/raw/production/Raw_2018_Production/"))

p18.exp <- exp.files2018 %>% 
  map_df(~ read_sheets_prod18exp(.)) %>% 
  clean_names() %>% 
  mutate(source = str_remove(file_name, "./data/raw/production/Raw_2018_Production/"))


```

```{r, eval=FALSE}
### qa
p18.exp %>% 
  select(!contains("x")) %>% 
  visdat::vis_dat() +
  labs(title = "2018 raw production data") +
  coord_flip()

p18.exp %>% 
  filter(!is.na(sl)) %>% 
  tabyl(source) %>% 
  gt() %>% 
  tab_header(title = "2018 production data", subtitle = "Raw input files: exp")

p18.obs %>% 
  filter(!is.na(sl)) %>% 
  tabyl(source) %>% 
  gt() %>% 
  tab_header(title = "2018 production data", subtitle = "Raw input files: obs")

## each of the sites has data entered into a separate, single-tabbed excel file

```


```{r}
### work with the combined obj and exp import
p18 <- p18 %>% 
  janitor::clean_names() %>% 
  mutate(yr = 2018) %>%
  mutate(yr = as.character(yr)) %>% 
  mutate(season = "fall")

```

```{r, eval=FALSE}
p18 %>% 
  tabyl(source) %>% 
  gt() %>% 
  tab_header(title = "2018 production data", subtitle = "Raw input files")

```


```{r}
## 
# eliminate  columns without data. I don't know if date is important (yr and season is already attributed). It's a mess, so just eliminating
p18 <- p18 %>% 
  select(!c(x129,date,date_year_month_day))

```

```{r, eval = TRUE}

# specific fix
p18 <- p18 %>% 
  mutate(site = case_when(is.na(site) & wilid == "408" ~ "elk",
                           TRUE ~ site))

```

```{r}
## Select subset and type cast
p18sel <- p18 %>% 
  select(yr, season, spp, site, plot, wilid, pl_ht_cm, source) %>%
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>% 
  distinct()

```
# >
```{r}
# Not all plots are attributed with plot type
# Replace obs for NA
p18sel %>% filter(is.na(plot))


p18sel <- p18sel %>% 
  mutate(plot = if_else(condition = is.na(plot), true = "obs", plot)) 

p18sel <- p18sel %>% 
  filter(!is.na(site)) # NA for site are also na for ht so are eliminated
```


```{r}
p18sel %>%
  visdat::vis_miss() +
  labs(caption = "Source: 2018 production data(many source xlsx files)")

```




```{r}
### missing willow id
p18sel %>% 
  filter(is.na(wilid)) %>% 
  gt() %>% 
  tab_header(title = "2018 production files: missing 'wilid'")

```


```{r}
p18sel %>% 
  tabyl(source) %>% 
  gt() %>% 
  tab_header(title = "2018 production data", subtitle = "Raw input files (numerous)")

```


```{r}
# p18sel %>% 
#   ggTile_yr_season_site() +
#   labs(caption = "p18sel")

p18sel %>% 
  ggTile_yr_season_plot() +
  labs(caption = "p18sel")

```

## 2018 Utilization

```{r}
ufiles2018 <- fs::dir_ls("./data/raw/utilization/Raw_2018_Utilization", recurse = FALSE, glob = "*.xlsx")
# view(ufiles2018)
## single sheet
# u18 <- ufiles2018 %>%
#   map_df( ~ read_excel(path = .x,trim_ws = TRUE, skip = 1, col_types = "text"))

## add code to read multitabs...
# u18 <- ufiles2018 %>% 
#   map_df(~ read_sheets(.))

u18 <- ufiles2018 %>% 
  map_df(~ read_sheets_util(.))

# map_df( ~ read_excel(path = .x,trim_ws = TRUE, skip = 1, col_types = "text", range = "A2:DV46"))
# # note use of specific range and skipping of column

u18 <- u18 %>% 
  janitor::clean_names() %>% 
  mutate(yr = "2018") %>% 
  mutate(season = "spring") 

u18 <- u18 %>%
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) 

u18 %>% 
  tabyl(plot, site) %>% 
  datatable()

```

```{r}
## whack a mole.
# change lb41 to lb4
u18 <- u18 %>% 
  mutate(site = case_when(site == "lb41" ~ "lb4",
                          TRUE ~ site)) 
```

```{r}
u18 %>% 
  names() %>%
  tibble::enframe(name = NULL) %>% 
  gt() %>% 
  tab_header(title = "variable names")

## WB?
## BIG fix: pl_ht and pl_ht_cm
## willid and wilid
u18 %>%
  select(file_name,contains("cm")) %>% 
  distinct()

u18 %>% 
  distinct(site) %>% gt()

```

```{r}
# u18 %>% names.dt()
## !!!
u18 <- u18 %>% 
  mutate(height_cm = as.numeric(height_cm)) %>% 
  mutate(ht_cm = as.numeric(ht_cm))
  
u18 %>% filter(is.na(pl_ht_cm) & !is.na(height_cm)) %>% 
# & is.na(ht_cm))
  select(pl_ht_cm, height_cm, ht_cm) %>% 
  glimpse()

u18 <- u18 %>% 
  mutate(pl_ht_cm = if_else(condition = is.na(pl_ht_cm),true = height_cm,false = pl_ht_cm)) %>% 
  mutate(pl_ht_cm = if_else(condition = is.na(pl_ht_cm),true = ht_cm, false = pl_ht_cm))

u18 <- u18 %>% 
  mutate(wilid = case_when(is.na(wilid) ~ willid,
                              is.na(willid)~ wilid,
                              TRUE ~ wilid))%>% 
  filter(!is.na(pl_ht_cm)) %>%
  select(-c(height_cm,willid))

### select subset of columns: for height
u18.ht <- u18 %>%
  # glimpse()
  select(c('file_name','yr','season', 'plot','site','wilid','pl_ht_cm')) %>% 
  distinct()

```

```{r}
### 2018 crescent
u18.ht %>% 
  filter(!is.na(pl_ht_cm)) %>% 
  # distinct(site)
  filter(site == "cresc") %>% 
  select(wilid, site, plot, pl_ht_cm) 
```

```{r}

# u18.ht %>% 
#   distinct(pl_ht_cm) %>% 
#   View() # some rows have na entred as text

# rename, reclass
u18.ht <- u18.ht %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>% 
  mutate(plot = if_else(condition = is.na(plot), true = "obs", plot))

# clean
u18.ht <- u18.ht %>% 
  distinct()

# more cleaning
u18.ht <- u18.ht %>% 
  filter(!is.na(pl_ht_cm)) %>% 
  filter(wilid != "?") %>% 
  mutate(wilid = str_remove(.$wilid, ".0")) # remove the ".0" part of strings

```

```{r}
##########
## Extract per wilid height from population of stid observations
##########
# u18.ht <- u18.ht %>% 
#   filter(!is.na(pl_ht_cm)) %>% 
#   group_by(yr, wilid, site, plot, season) %>% 
#   summarise(pl_ht_cm = max(pl_ht_cm)) %>% 
#   ungroup()

u18.ht <- u18.ht %>% 
  filter(!is.na(pl_ht_cm)) %>% 
  distinct()

u18.ht <- u18.ht %>% 
  mutate(yr = as.character(yr)) %>% 
  mutate(wilid = case_when(wilid == "0316" ~ "316",
                           TRUE ~ wilid)) 

u18.ht <- u18.ht %>% 
  filter(!is.na(wilid)) %>%
  filter(!wilid == "") %>% 
  mutate(wilid = str_remove(.$wilid, ".0"))


u18.ht %>% distinct(wilid) %>% datatable()

u18.ht %>% 
  tabyl(site, plot) %>% 
  datatable()

## where is wb???
u18.ht %>% 
  visdat::vis_dat()

```

```{r}
u18.ht %>% 
  ggTile_yr_season_plot()
```

No 'cx' for 2018 spring

```{r, eval = TRUE}
#### Diagnostic plots
u18.ht %>%
  # visdat::vis_guess() + 
  # visdat::vis_miss() +
  visdat::vis_dat() +
  theme_classic() +
  labs(title = "2018 Spring Height", subtitle = "Missing map for qa/qc") +
  scale_fill_viridis(discrete = TRUE, option = "D") +
  coord_flip()

```

```{r}

u18.plotly.exp <- u18.ht %>% 
  filter(plot != "obs") %>% 
  filter(pl_ht_cm <550) %>% 
  mutate(yr = as.character(yr)) %>% 
  ggplot(aes(site, pl_ht_cm)) +
  # geom_boxplot(aes(color = site)) +
  # geom_boxplot(aes(fill = plot)) +
  geom_jitter(alpha = .5, aes(shape = plot, color = plot)) +
  labs(y = "Height (cm)", title = "2018 Spring Data")
plotly::ggplotly(u18.plotly.exp)  

```

```{r}
# plot facet
u18.plotly.exp.bx <- u18.ht %>% 
  filter(plot != "obs") %>% 
  filter(pl_ht_cm <550) %>% 
  mutate(yr = as.character(yr)) %>% 
  ggplot(aes(plot, pl_ht_cm)) +
  # geom_boxplot(aes(color = site)) +
  geom_boxplot(aes(fill = plot)) +
  geom_jitter(alpha = .8, aes(shape = plot),color = "black") +
  facet_wrap(~site) +
  labs(y = "Height (cm)", title = "2018 Spring Data")
plotly::ggplotly(u18.plotly.exp.bx)  

```

```{r, fig.height=9}
# text plot
u18.plotly.exp.tx<- u18.ht %>% 
  filter(plot != "obs") %>% 
  filter(pl_ht_cm <550) %>% 
  mutate(yr = as.character(yr)) %>% 
  ggplot(aes(plot, pl_ht_cm)) +
  # geom_boxplot(aes(color = site)) +
  # geom_boxplot(aes(fill = plot)) +
  # geom_jitter(alpha = .8, aes(shape = plot),color = "black") +
  geom_text_repel(aes(label = wilid), size = 2.25) +
  geom_point(color = "blue") +
  # facet_wrap(~site) +
  facet_grid(~site) +
  labs(y = "Height (cm)", title = "2018 Spring Data")
u18.plotly.exp.tx

```

```{r}
u18.plotly.obs <- u18.ht %>% 
  filter(plot == "obs") %>% 
  mutate(yr = as.character(yr)) %>% 
  ggplot(aes(site, pl_ht_cm)) +
  # geom_boxplot(aes(color = site)) +
  geom_jitter(alpha = .5, aes(shape = plot, color = plot)) +
  labs(y = "Height (cm)", title = "2018 Spring Data")
plotly::ggplotly(u18.plotly.obs) 
```

## 2019 Production

```{r}
# updated as of 20191203
files2019 <- fs::dir_ls("./data/raw/production/Raw_2019_Production", recurse = FALSE, glob = "*.xlsx")

p19 <- files2019 %>% 
  map_df(~ read_sheets_prod(.))

p19 <- p19 %>% 
  janitor::clean_names() %>% 
  mutate(yr = '2019') %>% 
  mutate(season = "fall")

p19 %>%
  # View()
  # filter(!is.na(date))%>%
  distinct(file_name, sheet_name) %>%
  gt() %>% 
  tab_header(title = "2019 Fall Production Files and Sheets")

```

```{r}
## Select subset and type cast
p19sel <- p19 %>% 
  select(file_name, sheet_name, spp, yr, season, site, plot, wilid, pl_ht_cm) 

p19sel <- p19sel %>% 
  mutate(plot = if_else(condition = (is.na(plot)), true = "obs", plot)) 

p19sel <- p19sel %>% 
  mutate(plot = case_when(plot == "1" ~ "obs",
                          plot == "2" ~ "obs",
                          plot == "3" ~ "obs",
                          TRUE ~ plot)) %>% 
  mutate(site = case_when(site == "Crystal" ~ "crystal",
                          site == "cresc" ~ "crescent",
                          TRUE ~ site))


# prep species lookup
lu.wilid.spp.v3 <- lu.wilid.spp.v2 %>% 
  separate(col = wilid_full, into = c("site","plot", "wilid"),sep = "-",remove = FALSE)


## create new spp lu
lu.wilid.spp.v3 <- p19sel %>% 
  select(file_name, site, yr, plot, wilid, spp) %>%
  rename(species = spp) %>% 
  filter(!is.na(wilid)) %>% 
  distinct() %>%
  mutate(wilid_full = paste0(site,"-",plot,"-", wilid)) %>% 
  # distinct(site, plot) %>% datatable()
  bind_rows(.,lu.wilid.spp.v3) %>% 
  select(site, wilid, species, wilid_full) %>% 
  distinct()

lu.wilid.spp.v3 %>% 
  filter(species == "NA") %>% 
  gt() %>% 
  tab_header(title = "missing species attributes", subtitle = "2019 production")

# p19sel <- p19sel %>% 
#   filter(!(is.na(site) & is.na(pl_ht_cm)))

p19sel %>% 
  # filter(plot == '3') %>%
  # datatable()
  tabyl(site, plot) %>% 
  gt()

```

```{r}
# eliminate duplicate rows 
p19sel <- p19sel %>%
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>% 
  filter(!is.na(pl_ht_cm)) %>% 
  select(yr, season, site, plot, wilid, pl_ht_cm) %>% 
  distinct() 

p19sel %>% 
  ggTile_yr_season_site() +
  labs(caption = "p19.sel")

p19sel %>% 
  ggTile_yr_season_plot() +
  labs(caption = "p19.sel")

```

```{r}
## combine 18 and 19
p18p19sel <- bind_rows(p18sel, p19sel)

# visdat::vis_miss(p18p19sel)

# make all lower
p18p19sel <- p18p19sel %>%  
  mutate_if(.predicate = is.character,.funs=tolower)

p18p19sel <- p18p19sel %>%
  filter(!is.na(pl_ht_cm))

```

```{r}
# remove records that have na for wilid
p18p19sel <- p18p19sel %>% 
  filter(!is.na(wilid)) %>%
  filter(!is.na(pl_ht_cm)) %>% 
  mutate(site = case_when(is.na(site) & wilid == "408" ~ "elk",
                          TRUE ~ site)) %>% 
  # select(-c(file_name, spp, sheet_name, stid)) %>%
  mutate(site = case_when(site == "lb41" ~ "lb4",
                          site == 'lost c' ~ "lostc",
                          site == "lost l" ~ "lostl",
                          TRUE ~ site)) %>% 
  distinct()

## summarize the willow height by wilid and year
# p1819.wilidht <- p18p19sel %>% 
#   group_by(yr, site, spp, wilid) %>% 
#   summarise(pl_ht_cm = max(pl_ht_cm)) %>% 
#   ungroup()

```

```{r}

p1819.wilidht <- p18p19sel  

# eliminate the na for plant height
p1819.wilidht <- p1819.wilidht %>%
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid))

## some plots
p1819.wilidht %>% ggTile_yr_season_site() +
  labs(caption = "p1819.wilidht")

p1819.wilidht %>% ggTile_yr_season_site2()  +
  labs(caption = "p1819.wilidht")

p1819.wilidht %>% ggTile_yr_season_plot()  +
  labs(caption = "p1819.wilidht")

```

```{r}
## Additional 18 data?
# 
# Here is another file containing data: Exp_2018_Production_Height_20181220.xls
# Seems to befall data, but not just for the exp sites. Not sure if it's comprehensive of what was collected in the Fall of 2018. 

ht18fa.add <- read_sheets_prod(file = "./data/raw/production/Exp_2018_Production_Height_20181220.xlsx") %>% 
  janitor::clean_names()

ht18fa.add <- ht18fa.add %>% 
  select(c(site,plot,wilid, pl_ht_cm, sheet_name)) %>% 
  mutate(season = "fall") %>% 
  mutate(yr = as.character("2018")) %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>% 
  mutate(site2 = paste0(site,"-",plot)) %>%
  mutate(wilid_full = paste0(site,"-",plot,"-",wilid)) %>% 
  filter(!is.na(pl_ht_cm)) %>% 
  filter(!is.na(site)) %>% 
  distinct()

```

```{r}
### a different height file that has 2018 height data entered.
ht18fa.add %>% 
  ggTile_yr_season_site2() +
  labs(caption = "ht18fa.add")

ht18fa.add %>% 
  ggTile_yr_season_site() +
  labs(caption = "ht18fa.add")

##
p1819.wilidht <- ht18fa.add %>% 
  select(-c(sheet_name)) %>% 
  bind_rows(.,p1819.wilidht) %>% 
  distinct()

# p1819.wilidht %>% ggTile_yr_season_site2()

p1819.wilidht %>% ggTile_yr_season_plot()
```

Not sure whther these duplicate records in the ht data assembled from the raw production files

```{r}
# inspect before rowbind

# ht18fa.add %>% 
#   glimpse()
# 
# p1819.wilidht %>% 
#   glimpse()

p1819.wilidht %>% ggTile_yr_season_plot +
  labs(caption = "p1819.wilidht")

```

## 2019 Utilization

```{r}
#### 
ufiles2019 <- fs::dir_ls("./data/raw/utilization/Raw_2019_Utilization", recurse = FALSE, glob = "*.xlsx")

# u19 <- ufiles2019 %>%
#   map_df( ~ read_excel(path = .x,trim_ws = TRUE, skip = 1, col_types = "text"))

## multitab
# u19 <- ufiles2019 %>% 
#   map_df(~ read_sheets(.))

u19 <- ufiles2019 %>% 
  map_df(~ read_sheets_util(.))

# map_df( ~ read_excel(path = .x,trim_ws = TRUE, skip = 1, col_types = "text", range = "A2:DV46"))
# # note use of specific range and skipping of column

u19 <- u19 %>% 
  janitor::clean_names() %>% 
  mutate(yr = 2019) %>% 
  mutate(season = "spring") 

u19 %>% 
  names() %>%
  tibble::enframe(name = NULL) %>%
  datatable(caption = "field names in raw utilization files")

u19 <- u19 %>%
  filter(!is.na(pl_ht_cm)) %>% 
  # mutate(pl_ht_cm = case_when(is.na(height_cm)~ ht_cm,
  #                             is.na(pl_ht_cm)~height_cm,
  #                             TRUE ~ pl_ht_cm)) %>%
  mutate(wilid = case_when(is.na(wilid) ~ willid,
                              is.na(willid)~ wilid,
                              TRUE ~ wilid))

### select subset of columns: for UTILIZATION
# u19.sel <- u19 %>%
#   # glimpse()
#   select(c('date','yr','season', 'site','wilid','stid', 'live_status', 'scheme','height_cm','spp','len_dia'), contains("sht"),contains("x"))

```

```{r}
### select subset of columns: for height
u19.ht <- u19 %>%
  # glimpse()
  select(c(file_name, sheet_name, yr, site, season, spp, plot, wilid, stid, pl_ht_cm))


u19.ht <- u19.ht %>% 
  mutate(site = case_when(is.na(site) & wilid == "408" ~ "elk",
                          TRUE ~ site)) %>% 
  mutate(site = case_when(site == "lb41" ~ "lb4",
                          site == 'lost c' ~ "lostc",
                          site == "lost l" ~ "lostl",
                          TRUE ~ site)) %>% 
  mutate(wilid = case_when(wilid == '0316' ~ "316",
                           TRUE ~ wilid)) %>% 
  distinct() %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>% 
  filter(!is.na(pl_ht_cm))

u19.ht %>% ggTile_yr_season_site()
```

```{r}
# rename, reclass
u19.ht <- u19.ht %>% 
  # filter(is.na(site))
  # mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>% 
  mutate(plot = if_else(condition = is.na(plot), true = "obs", plot))

# clean
u19.ht <- u19.ht %>% 
  # filter(!(is.na(willid) & is.na(wilid))) %>% 
  distinct()

# more cleaning
u19.ht <- u19.ht %>% 
  filter(!is.na(pl_ht_cm)) %>%
  filter(wilid != "na")

visdat::vis_miss(u19.ht)

```

```{r}
u19 %>% 
  visdat::vis_miss() +
  theme(axis.text.x=element_blank()) +
  labs(caption = "Data structure of raw utilization files")

```

```{r}
# ##########
# ## calculate the per wilid height from population of stid
# ##########
# u19.ht <- u19.ht %>% 
#   filter(!is.na(pl_ht_cm)) %>% 
#   group_by(yr, wilid, site, plot, season) %>% 
#   summarise(pl_ht_cm = max(pl_ht_cm)) %>% 
#   ungroup()

# enforce lower case
u19.ht <- u19.ht %>%
  mutate_if(.predicate = is.character,.funs=tolower)

u19.ht <- u19.ht %>% 
  select(-c(file_name, sheet_name, stid)) %>% 
  distinct()

```

#### update species lu

```{r}

## spp lu update
# prep species lookup
lu.wilid.spp.v3

## create new spp lu
lu.wilid.spp.v3 <- u19.ht %>% 
  select(site, yr, plot, wilid, spp) %>%
  rename(species = spp) %>% 
  filter(!is.na(wilid)) %>% 
  distinct() %>%
  mutate(wilid_full = paste0(site,"-",plot,"-", wilid)) %>% 
  # distinct(site, plot) %>% datatable()
  bind_rows(.,lu.wilid.spp.v3) %>% 
  select(site, wilid, species, wilid_full) %>% 
  distinct()

```

```{r, eval = FALSE}
u19.ht %>%
  visdat::vis_miss(cluster = TRUE) +
  theme_classic() +
  labs(title = "2019 Spring Height", subtitle = "Missing map for qa/qc") +
  scale_fill_viridis(discrete = TRUE) +
  coord_flip()

# plotly::ggplotly()
```

```{r}
u19.ht <- u19.ht %>% 
  distinct() %>% 
  mutate(site = case_when(wilid == "415" & is.na(site) ~ "ghole",
                      wilid == "418" & is.na(site) ~ "ghole",
                      wilid == "449" & is.na(site) ~ "glen",
                      wilid == "452" & is.na(site) ~ "glen",
                      TRUE ~ site))

## adress bad plots
u19.ht <- u19.ht %>% 
  # distinct(plot)
  mutate(plot = case_when(plot == "1" ~ "obs",
                          plot == "2" ~ "obs",
                          TRUE ~ plot))

u19.ht %>% ggTile_yr_season_site() + labs(caption = "u19.ht; cleaned sites")

u19.ht %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  ggTile_yr_season_site2() + 
  labs(caption = 'u19.ht; cleaned sites')

```

```{r}
u19.ht %>% 
  gt_tally_site_plot_yr_season()

```

```{r}
u19.ht %>% 
  mutate(yr = as.character(yr)) %>%
  ggPoint_yr_site_season_plot()
```

## Combine and QC

## Tom missing records

File provided by TH: *2001-2015time_series.csv*

```{r}

ts01.15 <- read_csv("./data/Tom_missingrecs/2001-2015time_series.csv")

gt_ts01.15t <- read_csv("data/Tom_missingrecs/2001-gt.2015time_series.csv")

th16.19 <- read_csv("data/Tom_missingrecs/2016-2019time_series.csv")

```

```{r}
## lots of dead rows imported b/c of the conservative read of the tabs in the excel workbooks. Eliminate these...
p18 <- p18 %>% 
  filter(!is.na(wilid) & !is.na(stid))

p18 %>%
  # filter(is.na(plot)) %>%
  tabyl(site) %>% 
  gt() 
# assume NA for plot == obs

```

```{r}
### Combine 18 and 19 spring data
# type convert
u19.ht <- u19.ht %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>% 
  mutate(yr = as.character(yr)) %>% 
  filter(!is.na(pl_ht_cm))

# row bind in 18
u18u19.ht <- bind_rows(u18.ht,u19.ht)

## correct error with plot = "3" to obs
u18u19.ht <- u18u19.ht %>%
  mutate(plot = case_when(plot == 3 ~ "obs",
                          TRUE ~ plot)) %>%
  mutate(site = case_when(site == "lost c" ~ "lostc",
                          site == "lost l" ~ "lostl",
         TRUE ~ site)) %>% 
  mutate(wilid_full = paste0(site,"-",plot,"-", wilid)) 

## clean: on record at elk has a ht meassurement but no site attribute...
u18u19.ht <- u18u19.ht %>%
  mutate(site = case_when(is.na(site) & file_name == "./data/raw/utilization/Raw_2018_Utilization/elk_2018_utilization.xlsx" ~ "elk",
TRUE ~ site)) 

## add column for tracking
# Raw_2018_Utilization/elk_2018_utilization.xlsx

u18u19.ht %>% 
  tabyl(site)

u18u19.ht %>% 
  tabyl(plot)

# plot
u18u19.ht %>% 
  mutate(site2 = paste0(site,"-",plot)) %>%
  ggTile_yr_season_site2(.) +
  labs(title = "Count of records, 2018 and 2019 utilization files", caption = "u18u19.ht")

```

```{r, eval = FALSE}
u18u19.ht %>% 
  # filter((wilid %in% c("415", "418", "449", "449", "452")) & is.na(site)) %>%
  # filter(wilid %in% c("415", "418", "449", "449", "452")) %>%
  distinct(wilid, site, yr) %>% 
  arrange(wilid)

# u18u19.ht <- u18u19.ht %>% 
#   filter(!is.na(site))

```

```{r, eval=FALSE}
## qa/qc
p1819.wilidht %>% 
  ggPoint_yr_site_season_plot

```

```{r}
# fix na for plot
p1819.wilidht <- p1819.wilidht %>%
  mutate(plot = if_else(condition = is.na(plot),true = "obs", plot)) %>%
  mutate(site2 = paste0(site,"-",plot))

p1819.wilidht <- p1819.wilidht %>% 
  mutate(season = "fall")

```

```{r}

## adress missing sites
# u18u19.ht <- u18u19.ht %>% 
#   mutate(site = case_when(is.na(site.x) ~ site.y,
#                           # is.na(site.y) ~ site.x,
#                           TRUE ~ site.x)) %>%
#   select(-c(site.x, site.y)) %>%
#   filter(!is.na(site))

u18u19.ht <- u18u19.ht %>% 
  mutate(plot = if_else(condition = plot == "1",true = "obs","obs")) ## not sure why, but error with plot...

# u18u19.ht %>% visdat::vis_miss()

```

```{r}
### Combine spring and fall for 2018 and 2019
##
p1819.wilidht <- p1819.wilidht %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm))

ht.all18.19 <- bind_rows(p1819.wilidht,u18u19.ht)

## eliminate NA fpr plant ht
ht.all18.19 <- ht.all18.19 %>% 
  filter(!is.na(pl_ht_cm)) %>% 
  filter(!is.na(wilid))

```

```{r}
## ### 2018 2019
# ### try to add spp info to the 18 and 19 
# wilid.spp.lu2 <- read_csv("./data/wilid_spp._lu_v2.csv") %>% 
#   select(c(species, wilid_full))

## address sites without proper site attribution
ht.all18.19 <- ht.all18.19 %>%
  mutate(site = case_when(wilid == "415" & is.na(site) ~ "ghole",
                      wilid == "418" & is.na(site) ~ "ghole",
                      wilid == "449" & is.na(site) ~ "glen",
                      wilid == "452" & is.na(site) ~ "glen",
                      TRUE ~ site)) %>%
  # select(-file_name) %>% 
  mutate(site = case_when(site == "lb41" ~ "lb4",
                          site == 'lost c' ~ "lostc",
                          site == "lost l" ~ "lostl",
                          TRUE ~ site)) %>% 
  mutate(site2 = paste0(site,"-", plot)) %>% 
  mutate(wilid_full = paste0(site2,"-", wilid))

ht.all18.19 <- ht.all18.19 %>% 
  mutate(wilid = case_when(wilid == "" & site == "lb5" ~ '1070',
                           TRUE ~ wilid))

ht.all18.19 <- ht.all18.19 %>% 
  distinct()

ht.all18.19 %>% 
  visdat::vis_miss()

```

Missing data: Yancy fall 2018

```{r}

ht.all18.19 <- ht.all18.19 %>% 
  mutate(site = case_when(site == 'cresc'~'crescent',
                          TRUE ~ site))

ht.all18.19 %>% 
  filter(plot != "obs") %>% 
  ggTile_yr_season_site2() +
  labs(title = "Exp sites", caption = 'ht.all18.19')

ht.all18.19 %>% 
  filter(!(plot != "obs")) %>% 
  ggTile_yr_season_site2() +
  labs(title = "Obs sites", caption = 'ht.all18.19')

```

```{r}
### Bring in 16 17
## 16
p16.ht <- p16.ht %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(yr = as.character("2016")) %>% 
  mutate(season = "fall")

p16.ht %>% 
  tabyl(site2) %>% 
  gt() %>% 
  tab_header(title = "n distinct fall 2016 site2")

## correct site id data entry error
p17.ht <- p17.ht %>% 
  mutate(site = if_else(site == "wn","wb",site)) %>%   mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(yr = as.character("2017")) %>% 
  mutate(season = "fall")

####
p16.ht <- p16.ht %>%
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) 

# p17.ht <- p17.ht %>% 
#  select(-species)

p17.ht %>% 
  tabyl(site2, season) %>% 
  arrange(site2) %>% 
  gt() %>% 
  tab_header(title = "n distinct fall 2017 site2")

# p17.ht %>%
#   filter(wilid == "236") %>% filter(site2 == "wn-cx") ## corected

p17.ht <- p17.ht %>% 
  mutate(wilid = as.integer(wilid))

p16p17 <- bind_rows(p16.ht,p17.ht)

p16p17 %>% 
  visdat::vis_dat() +
  labs(caption = "p16p17")

# Eliminate duplicates
# p16p17 %>% 
#   get_dupes() %>% View()
p16p17 <- p16p17 %>% 
  distinct() 
```

```{r}
p16p17 %>% 
  group_by(yr, season, site) %>%
  summarise(n= n()) %>%
  ggplot(aes(yr, site)) +
  geom_tile(aes(fill = n), color = 'black') +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 7)) +
  theme(axis.text.y = element_text(size = 9)) +
  facet_wrap(~season) +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 9)) +
  theme(axis.text.y = element_text(size = 7)) +
  facet_wrap(~season) +
  scale_fill_viridis(option = "D") +
  labs(caption = "Combined 2016 and 2017 fall height")
```

```{r}
## spring 16 17
u16.ht <- u16.ht %>%
  mutate(season = "spring") %>% 
  mutate(yr = as.character("2016")) %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm))

u17.ht <- u17.ht %>% 
  mutate(season = "spring") %>% 
  mutate(yr = as.character("2017"))

```

## Supplemental files from LM

## File: 2016_2017_Data_Entry_Format_WillowHt_2001_2015.csv

Another data file. Not sure how complete it is or how it was produced. LM created the file I believe from raw field forms(?)

*File: 2016_2017_Data_Entry_Format_WillowHt_2001_2015.csv*

```{r}

ht_lm_entry <- read_csv("./data/raw/ht/2016_2017_Data_Entry_Format_WillowHt_2001_2015.csv")

ht_lm_entry <- ht_lm_entry %>% 
  janitor::clean_names() %>% 
  select(-id) %>% 
  rename(wilid = willid) %>% 
  mutate(plot = case_when(dam == "1" & browse == "1" ~ "dc",
                          dam == "1" & browse == "0" ~ "dx",
                          dam == "0" & browse == "1" ~ "cc",
                          dam == "0" & browse == "0" ~ "cx",
                          TRUE ~ "obs"))

ht_lm_entry <- ht_lm_entry %>%
  rename(yr = year) %>%
  rename(pl_ht_cm = plantht) %>%
  mutate_if(.predicate = is.numeric,.funs = as.character) %>%
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) 

```

```{r}
ht_lm_entry %>%
  mutate(site2 = paste0(site,"-",plot)) %>% 
  group_by(yr, season, site2) %>% 
  tally() %>% 
  # filter(yr == "2017") %>% 
  # filter(season != "fall") %>%
  ggplot(aes(season, site2)) +
  geom_tile(aes(fill = n)) +
  facet_wrap(~yr, ncol = ) +
  labs(caption = "raw: 2016_2017_Data_Entry_Format_WillowHt_2001_2015.csv")

```

```{r}

```

```{r, echo = TRUE}

ht_lm_entry %>%
  # filter(yr == "2016" | yr == '2017') %>%
  select(-c(exp,dam,browse)) %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid)) %>% 
  filter(is.na(pl_ht_cm)) %>% 
  gt() %>% 
  tab_header(title = "records missing height", subtitle = "2016_2017_Data_Entry_Format_WillowHt_2001_2015.csv")

### ignoring all the data 2015 and earlier as Tom says he has this. Also removing records with NA for pl_ht_cm:
ht_lm_entry_16_17 <- ht_lm_entry %>%
  filter(yr == "2016" | yr == '2017') %>%
  select(-c(exp,dam,browse)) %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid)) %>% 
  filter(!is.na(pl_ht_cm)) %>% 
  distinct() 

###
# u17.ht %>% 
#   distinct(site2) %>% 
#   gt() %>% 
#   tab_header(title = "distinct site")

### combine ht from 2017 utilization data with the file LM sent 
u17_lm16_17 <- u17.ht %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>% 
  bind_rows(., ht_lm_entry_16_17) %>% 
  distinct()


```


## Combine various 2016-2019 data objects

```{r}
#### bind rows p16, p17, all ht 18 19
# ht.all18.19 %>% distinct(wilid) %>% View()

ht.all18.19 <- ht.all18.19 %>% 
  mutate(wilid = as.integer(wilid))

#### bind rows p16, p17, all ht 18 19
p.16.17.18.19 <- ht.all18.19 %>%
  # rename(live_dead = live_status) %>% 
  mutate(yr = as.character(yr)) %>% 
  bind_rows(., p16p17) %>% 
  distinct()

## add in the other lewis ht entry file

u17_lm16_17 <- u17_lm16_17 %>% 
  mutate(wilid = as.integer(wilid))


p.16.17.18.19 <- p.16.17.18.19 %>% 
  bind_rows(., u17_lm16_17) %>%
  # select(-live_dead) %>%
  distinct()

# p.16.17.18.19 %>% 
#   visdat::vis_dat()

```

```{r}
p.16.17.18.19 %>% ggTile_yr_season_plot_alt1()
```

```{r}

u18u19.ht <- u18u19.ht %>% 
  mutate(yr = as.character(yr)) %>% 
  mutate(wilid = as.integer(wilid))

ht.all.16.17.18.19 <- p.16.17.18.19 %>% 
  bind_rows(., u18u19.ht) %>% 
  # select(-c(file_name,live_status)) %>% 
  distinct()

```

```{r}
ht.all.16.17.18.19 %>% 
  distinct()

```


### Combine with the ht extracted from prodx.csv and sprcagx.csv

Scraping these files for any data

```{r}

ht.all.16.17.18.19 <- ht.all.16.17.18.19 %>% 
  mutate(site = case_when(site == "halibuff" ~ "hailbuff",
                          site == "cresc" ~ 'crescent',
                          site == "crescent" ~ "crescent",
                          site == "lost l" ~ "lostl",
                          site == "lost c" ~ "lostc",
                          site == "xtal" ~ "crystal",
                          TRUE ~ site)) %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid))

ht.all.16.17.18.19 %>% 
  ggTile_yr_season_site2()

```

```{r}
ht.all.16.17.18.19 <- ht.all.16.17.18.19 %>% 
  mutate(site = case_when(wilid == "415" & is.na(site) ~ "ghole",
                      wilid == "418" & is.na(site) ~ "ghole",
                      wilid == "449" & is.na(site) ~ "glen",
                      wilid == "452" & is.na(site) ~ "glen",
                      TRUE ~ site)) 

ht.all.16.17.18.19 <- ht.all.16.17.18.19 %>%
  mutate(plot = case_when(site == "crescent" & plot == "cc" ~ "obs",
                          site == "cresc" & plot == "cc" ~ "obs",
                          site == "elk1" & plot == "cc" ~ "obs",
                          site == "elk2" & plot == "cc" ~ "obs",
                          site == "elk3" & plot == "cc" ~ "obs",
                          site == "elk4" & plot == "cc" ~ "obs",
                          site == "elk5" & plot == "cc" ~ "obs",
                          site == "ghole" & plot == "cc" ~ "obs",
                          site == "hailbuf" & plot == "cc" ~ "obs",
                          site == "lava" & plot == "cc" ~ "obs",
                          site == "lb" & plot == "cc" ~ "obs",
                          site == "lb1" & plot == "cc" ~ "obs",
                          site == "lb2" & plot == "cc" ~ "obs",
                          site == "lb3" & plot == "cc" ~ "obs",
                          site == "lb4" & plot == "cc" ~ "obs",
                          site == "lostc" & plot == "cc" ~ "obs",
                          site == "lostl" & plot == "cc" ~ "obs",
                          site == "oxbow" & plot == "cc" ~ "obs",
                          site == "rose" & plot == "cc" ~ "obs",
                          site == "slide" & plot == "cc" ~ "obs",
                          site == "wb1" & plot == "cc" ~ "obs",
                          site == "wb2" & plot == "cc" ~ "obs",
                          site == "wb4" & plot == "cc" ~ "obs",
                          site == "yancy" & plot == "cc" ~ "obs",
                          TRUE ~ plot
                          )) %>% 
  mutate(site2 = paste0(site,"-", plot)) %>% 
  mutate(wilid_full = paste0(site2,"-", wilid))

```

```{r}
ht.all.16.17.18.19 %>% 
  group_by(yr, season, site) %>%
  summarise(n= n()) %>%
  ggplot(aes(yr, site)) +
  geom_tile(aes(fill = n), color = 'black') +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 7)) +
  theme(axis.text.y = element_text(size = 9)) +
  facet_wrap(~season) +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 9)) +
  theme(axis.text.y = element_text(size = 7)) +
  facet_wrap(~season) +
  scale_fill_viridis(option = "C") +
  labs(caption = "Combined 2016 and 2017 fall height")
```

```{r}
# COMBINE prod and sprcag object with the ht.all from ht.all.16.17.18.19

# type convert wilid for rbind
prod.sprcag.ht <- prod.sprcag.ht %>% 
  mutate(wilid = as.integer(wilid))

# bind 
ht.all.01.19 <- ht.all.16.17.18.19 %>% 
  bind_rows(.,prod.sprcag.ht) %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid)) %>% 
  filter(!is.na(pl_ht_cm)) %>% 
  distinct()

# fix 
ht.all.16.17.18.19 %>% ggTile_yr_season_site2()



ht.all.01.19 %>% 
  visdat::vis_dat()

ht.all.01.19 %>% 
  ggTile_yr_season_site()

```

## File: 'MIssingWillowRecordsFromLewis_20201118.csv'

File of missing records sent to me by TH on 20201118. Not sure how it was produced. LM created the file I believe from raw field forms(?).

*File: MIssingWillowRecordsFromLewis_20201118.csv* (Note: I added the date to the end of the file that was sent to me. Not sure the date of origin)

```{r}
lm.20201118 <- read_csv("./data/raw/MIssingWillowRecordsFromLewis_20201118.csv") %>% 
  janitor::clean_names()

## There are a few notes in X11; combine with 'notes'
lm.20201118 <- lm.20201118 %>%
  mutate(notes = glue::glue("{notes} {x11}")) %>% 
  dplyr::select(-x11) %>% 
  mutate(notes = str_remove(notes,"NA"))

## a new variation on 'wilid': 'wildid'
## Also, in the file from LM (via TH), season is encoded differently than elsewhere.

lm.20201118 <- lm.20201118 %>%
  rename(wilid = wildid) %>% 
  rename(yr = year) %>% 
  mutate(season = case_when(season == "production" ~ "fall",
                            season == "utilization" ~ "spring",
                            TRUE ~ season))

```

```{r, eval=FALSE}
lm.20201118 %>%
  visdat::vis_dat() +
  labs(title = "File: MIssingWillowRecordsFromLewis_20201118.csv")

```

A variety of issues to address...

-   Missing data in various columns (e.g., site, height_cm)

-   Fields are names differently than other files

-   "season" is entered differently than other files

-   "LIVE" is inconsistently entered

```{r, eval=FALSE}
#### Dataframe summary for qa
print(dfSummary(lm.20201118, plain.ascii = FALSE, style = "grid", 
          graph.magnif = 0.75, valid.col = FALSE, tmp.img.dir = "/tmp"),
      method = "render")

## cross table: wilid x yr
# print(ctable(x = lm.20201118$wilid, y = lm.20201118$yr, prop = "r"),
#       method = "render")
```

```{r}
## there are cleaning issues. Missing site for one record with a ht measurement. Did detective work and determined that it is elk
lm.20201118 <- lm.20201118 %>%
  mutate(site = case_when(is.na(site) & wilid == 408 ~ "elk",
                          TRUE ~ site))

## add wilid_full
lm.20201118 <- lm.20201118 %>%
  mutate(wilid_full = paste0(site,"-",plot,"-",wilid)) 

```

```{r}
## some cleaning/renaming
lm.20201118 <- lm.20201118 %>%
  mutate(yr = as.character(yr)) %>% 
  mutate() %>% 
  mutate(wilid = as.integer(wilid)) %>% 
  rename(pl_ht_cm = height_cm)

```


```{r}
# File: MIssingWillowRecordsFromLewis_20201118.csv

lm.20201118 %>% 
  group_by(wilid_full, yr, season) %>%
  tally() %>% 
  ungroup() %>% 
  pivot_wider(names_from = season, values_from = n) %>% 
  datatable(rownames = FALSE, caption = "Count of records by season, File: MIssingWillowRecordsFromLewis_20201118.csv")

```

```{r, eval=FALSE}
lm.20201118 %>% 
  group_by(site, plot, wilid, yr, season) %>%
  tally() %>% 
  ungroup() %>% 
  pivot_wider(names_from = season, values_from = n) %>% 
  datatable(rownames = FALSE, caption = "Count of records by season, File: MIssingWillowRecordsFromLewis_20201118.csv")
```


```{r}
lm.20201118 %>% 
  distinct(live) %>% 
  gt() %>% 
  tab_header(title = "'live' is not consistenly attributed.", subtitle = "File: MIssingWillowRecordsFromLewis_20201118.csv")


```

*What is 'nd'? 
*Do you want to eliminate dead, missing, etc. plants?

```{r}
## create a lookup to compare with previously imported and munged data (e.g., prod files)
lm.20201118.wilid_full.lu <- lm.20201118 %>%
  distinct(wilid_full) 

```

```{r}
lm.20201118.wilid_full.lu %>% 
  gt() %>% 
  tab_header(title = "Distinct wilid_full", subtitle = "File: MIssingWillowRecordsFromLewis_20201118.csv")
```


```{r}
# Filtering joins filter rows from x based on the presence or absence of matches in y:
# 
# semi_join() return all rows from x with a match in y.
# 
# anti_join() return all rows from x without a match in y.

## use LU from missing records
# anti_join()

```

## Export 2016-2018 height



## Bind in records from DCC




```{r}

## bind in the dcc from raw files
dcc.ht.fa.spr <- dcc.ht.fa.spr %>%
  select(-c(species,wilid_full)) %>% 
  distinct()

dcc.ht.fa.spr %>% 
  visdat::vis_dat()

## 
# ht.all.16.17.18.19 %>% 
#   mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>% 
#   mutate(yr = as.character(yr))

## wilid type-convert for rbind 
dcc.ht.fa.spr <- dcc.ht.fa.spr %>% 
  mutate(wilid = as.integer(wilid))


dcc.16.17.18.19 <- ht.all.16.17.18.19 %>% 
  # select(-c(browse, exp, dam)) %>% 
  bind_rows(., dcc.ht.fa.spr) %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid)) %>% 
  filter(!is.na(pl_ht_cm))

dcc.16.17.18.19 <- dcc.16.17.18.19 %>% 
  distinct()

dcc.16.17.18.19 %>% 
  visdat::vis_dat() +
  labs(caption = 'dcc.16.17.18.19')

```

```{r}
dcc.16.17.18.19 %>%
  group_by(yr, season, site) %>%
  summarise(n= n()) %>%
  ggplot(aes(yr, site)) +
  geom_tile(aes(fill = n), color = 'black') +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 7)) +
  theme(axis.text.y = element_text(size = 9)) +
  facet_wrap(~season) +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 9)) +
  theme(axis.text.y = element_text(size = 7)) +
  facet_wrap(~season)

```

```{r}
dcc.16.17.18.19 <- dcc.16.17.18.19 %>% 
  mutate(wilid = as.integer(wilid))

## bind in other dcc object
dcc.16.17.18.19 <- ht.all18.19 %>% 
  bind_rows(.,dcc.16.17.18.19) %>%
  # select(-c(species, live_status)) %>% 
  distinct()

### select
dcc.16.17.18.19 <- dcc.16.17.18.19 %>% 
  select(c(yr, season, site, site2, wilid, wilid_full, pl_ht_cm, plot)) 

##
w_ht <- w_ht %>%
  # rename(plot = treat) %>%
  rename(yr = year) %>%   
  mutate(yr = as.character(yr)) %>% 
  mutate(wilid_full = paste0(site,"-",plot,"-", wilid))

w_ht <- w_ht %>% 
  select(c(yr, season, site, site2, wilid, wilid_full, pl_ht_cm, plot))%>% 
  mutate(wilid = as.character(wilid)) 

## join in spp
# w_ht <- left_join(w_ht, wilid.spp.lu2)

# FALL 18 and 19 plus DCC
# type convert
w_ht <- w_ht %>% 
  mutate(wilid = as.integer(wilid))

ht.combined <- dcc.16.17.18.19 %>% 
  mutate(pl_ht_cm = as.numeric(pl_ht_cm)) %>% 
  bind_rows(., w_ht) %>% 
  distinct()

## clean
ht.combined <- ht.combined %>% 
  mutate(site = case_when(site == "halibuff" ~ "hailbuff",
                          TRUE ~ site)) %>% 
  distinct()

```

```{r}
ht.combined %>%
  group_by(yr, season, site) %>%
  summarise(n= n()) %>%
  ggplot(aes(yr, site)) +
  geom_tile(aes(fill = n), color = 'black') +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 7)) +
  theme(axis.text.y = element_text(size = 7)) +
  facet_wrap(~season) +
  labs(caption = "Combined DCC, 16-19")

# dcc.ht.fa.spr %>%
#   group_by(yr, season, site) %>%
#   summarise(n= n()) %>%
#   ggplot(aes(yr, site)) +
#   geom_tile(aes(fill = n), color = 'black') +
#   theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 7)) +
#   theme(axis.text.y = element_text(size = 7)) +
#   facet_wrap(~season) +
#   labs(title = "DCC data")

```

```{r}
# valid?
# ht.combined <- ht.combined %>%
#   filter(pl_ht_cm < 700)

```

```{r, eval=FALSE}
# qa
ht.combined %>% 
  group_by(site2, season, yr) %>% 
  tally() %>% 
  ggplot(aes(yr, site2)) +
  geom_tile(aes(fill = n)) +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 7)) +
  theme(axis.text.y = element_text(size = 7)) +
  facet_wrap(~season) +
  labs(caption = "Combined DCC, 16-19")
```

```{r}
### the missing spring dcc data
# note: this is supplanted by code that pulls in 
# both fa and spr data from raw dcc files

# nsf.ht.spr <- nsf.ht.spr %>%
#   mutate(yr = as.character(yr), wilid = as.character(wilid)) %>%
#   mutate(season = "spring") %>%
#   mutate(site2 = paste0(site,"-",plot))
# nsf.ht.spr <- left_join(nsf.ht.spr, plot.lu, by = "plot")
# nsf.ht.spr %>% filter(is.na(species))

nsf.ht.fa.expobs <- nsf.ht.fa.expobs %>% 
  mutate(yr = as.character(yr)) %>% 
  mutate(wilid = as.character(wilid)) %>% 
  mutate(wilid_full = paste0(site,"-",plot,"wilid")) %>% 
  mutate(site2 = paste0(site,"-",plot))

```

```{r}

nsf.ht.fa.expobs <- nsf.ht.fa.expobs %>% 
  mutate(wilid = as.integer(wilid))

# ht.combined <- bind_rows(ht.combined, nsf.ht spring and fall) 
ht.combined <- bind_rows(ht.combined, nsf.ht.fa.expobs) %>% 
  distinct()


## create species lookup table 
# ht.combined %>% 
#   distinct(species) %>% 
#   write_csv("./data/lu_species_code.csv")
  
# clean
ht.combined <- ht.combined %>%
  # mutate(species = case_when(species == 'beb' ~ "bebb",
  #                            species == 'drumm' ~ "drum",
  #                            species == 'boothii' ~ "booth",
  #                            species == 'boothi' ~ "booth",
  #                            species == 'lassiandra' ~ "lass",
  #                            TRUE ~ species)) %>% 
  # filter(species != "?") %>%
  mutate(plot = if_else(is.na(plot),"obs",plot)) %>%   # filter(!(wilid == "570" & pl_ht_cm == 283)) %>% 
  filter(!is.na(site)) %>% 
  mutate(site = if_else(site == "crescent","cresc", site)) %>%
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site,"-",plot,"-",wilid))


```

```{r}

# ## find missing data
ht.combined %>%
  group_by(yr, season, site) %>%
  summarise(n= n()) %>%
  ggplot(aes(yr, site)) +
  geom_tile(aes(fill = n), color = 'black') +
  facet_wrap(~season) +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 7)) +
  theme(axis.text.y = element_text(size = 7)) +
  facet_wrap(~season)
# ggsave("./output/output4qaqc/height_heatmap20191204 1937.pdf", width = 11, height = 8.5)

```

```{r}
#clean sites
ht.combined <- ht.combined %>% 
  mutate(site = case_when(site == "lb41" ~ "lb4",
                          site == 'lost c' ~ "lostc",
                          site == "lost l" ~ "lostl",
                          TRUE ~ site)) %>% 
  mutate(site = case_when(site == "halibuff" ~ "hailbuff",
                          site == "cresc" ~ 'crescent',
                          site == "crescent" ~ "crescent",
                          site == "lost l" ~ "lostl",
                          site == "lost c" ~ "lostc",
                          site == "xtal" ~ "crystal",
                          TRUE ~ site))

# ht.combined <- ht.combined %>% 
#   mutate(pl_ht_cm = round(pl_ht_cm,0))

ht.combined <- ht.combined %>% 
  mutate(wilid = as.character(wilid)) %>% 
  filter(wilid != "0") %>% 
  distinct() 

ht.combined <- ht.combined %>% 
  mutate(site = case_when(site == "hailbuff"~ "hailbuf",
                          TRUE ~ site))

```

```{r}
#
ht.combined %>%
  group_by(yr, season, site2) %>%
  summarise(n= n()) %>%
  ggplot(aes(yr, site2)) +
  geom_tile(aes(fill = n), color = 'black') +
  facet_wrap(~season) +
  theme(axis.text.x = element_text(angle = 55, hjust = 1, size = 7)) +
  theme(axis.text.y = element_text(size = 7)) +
  facet_wrap(~season)

```

```{r}
## correct the height
ht.combined <- ht.combined %>%
  mutate(pl_ht_cm = if_else(pl_ht_cm == 1362, 362, pl_ht_cm)) %>% 
  mutate(pl_ht_cm = if_else(pl_ht_cm == 2877, 287, pl_ht_cm))

# these should be corrected in raw files?
```

### Rebind spring

```{r}
#### Site problems 18 19 spring
####### 
u18u19.ht <- u18u19.ht %>% 
  mutate(yr = as.character(yr)) %>% 
  select(-file_name)

u18u19.ht %>%
  distinct() %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  ggTile_yr_season_site2()


ht.combined <- ht.combined %>% 
  select(-c(species, wilid_full, site2)) %>% 
  distinct()
  
## bind in the spring19

ht.combined <- ht.combined %>% 
  bind_rows(., u19.ht) %>% 
  distinct()

ht.combined %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  ggTile_yr_season_site2()

#### 18
u18.ht <- u18.ht %>% 
  select(-c(file_name))

ht.combined <- ht.combined %>% 
  bind_rows(., u18.ht) %>% 
  distinct()

# ht.combined <- ht.combined %>% 
#   bind_rows(., u18u19.ht) %>% 
#   distinct()

ht.combined <- ht.combined %>%
  mutate(site = case_when(site == "cresc" ~ "crescent",
                          site == "hailbuff" ~ "hailbuf",
                          TRUE ~ site)) %>% 
  mutate(site = case_when(site == "lb41" ~ "lb4",
                          site == 'lost c' ~ "lostc",
                          site == "lost l" ~ "lostl",
                          TRUE ~ site)) %>% 
  mutate(plot = case_when(site == "hailbuf" & plot == "cc" ~ "obs",
                          site == "glen" & plot == "cc" ~ "obs",
                          TRUE ~ plot)) %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid))

ht.combined %>% ggTile_yr_season_site2()

```

```{r}
## more site name cleanup
ht.combined <- ht.combined %>% 
  select(yr, wilid, site, site2, season, plot, pl_ht_cm, wilid_full) %>%
  distinct()

ht.combined %>% ggTile_yr_season_site2()

```

```{r}
ht.combined <- ht.combined %>% 
  # select(-species) %>% 
  mutate(plot = case_when(site == "crescent" & plot == "cc" ~ "obs",
                          site == "cresc" & plot == "cc" ~ "obs",
                          site == "elk1" & plot == "cc" ~ "obs",
                          site == "elk2" & plot == "cc" ~ "obs",
                          site == "elk3" & plot == "cc" ~ "obs",
                          site == "elk4" & plot == "cc" ~ "obs",
                          site == "elk5" & plot == "cc" ~ "obs",
                          site == "ghole" & plot == "cc" ~ "obs",
                          site == "hailbuf" & plot == "cc" ~ "obs",
                          site == "lava" & plot == "cc" ~ "obs",
                          site == "lb" & plot == "cc" ~ "obs",
                          site == "lb1" & plot == "cc" ~ "obs",
                          site == "lb2" & plot == "cc" ~ "obs",
                          site == "lb3" & plot == "cc" ~ "obs",
                          site == "lb4" & plot == "cc" ~ "obs",
                          site == "lostc" & plot == "cc" ~ "obs",
                          site == "lostl" & plot == "cc" ~ "obs",
                          site == "oxbow" & plot == "cc" ~ "obs",
                          site == "rose" & plot == "cc" ~ "obs",
                          site == "slide" & plot == "cc" ~ "obs",
                          site == "wb1" & plot == "cc" ~ "obs",
                          site == "wb2" & plot == "cc" ~ "obs",
                          site == "wb4" & plot == "cc" ~ "obs",
                          site == "yancy" & plot == "cc" ~ "obs",
                          TRUE ~ plot
                          )) %>%
  mutate(plot = case_when(site == "glen" ~ "obs",
                          site == "hailbuff" ~ "obs",
                          TRUE ~ plot)) %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid))

```

```{r, eval = FALSE}
## try to retireve the 18 spring data that went missing 
ht.combined  <- ht.combined %>% 
  select(-c(site2, wilid_full)) %>% 
  bind_rows(., u18.ht) %>% 
  distinct() %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid))

ht.combined %>% 
  filter(plot != "obs") %>% 
  ggTile_yr_season_site2()

## try to retireve the 19 spring data that went missing 
ht.combined  <- ht.combined %>% 
  select(-c(site2, wilid_full)) %>% 
  bind_rows(., u19.ht) %>% 
  distinct() %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid))

ht.combined %>% 
  # filter(plot != "obs") %>% 
  ggTile_yr_season_site2()

## fix obvious site issues
ht.combined <- ht.combined %>% 
  select(-c(site2, wilid_full)) %>%
  mutate(site = case_when(site == "cresc" ~ "crescent",
                          site == "lost l" ~ "lostl",
                          site == "lost c" ~ "lostc",
                          site == "hailbuf" ~"hailbuff",
                          TRUE ~ site)) %>% 
  distinct() %>%   
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid))

ht.combined %>% ggTile_yr_season_site2()

```

```{r}
## what happended to 18 spr cresc?
u18.ht %>% 
  # distinct(site) %>% 
  filter(site == "cresc")

```

```{r}
## 
ht.combined %>% 
  ggTile_yr_season_site2()

```

### Data checks

### Specific issues Tom found regarding wb out of range data

```{r}
## Fa raw DCC. Note error in wilid=158, 2010 
 read_delim("./data/NSF_DataArchive20180208/Plant_level_fall_current_annual_growth_and_height_on_experimental_plot/willows-plantCAGfall2002-2015.txt", delim = " ") %>% 
  select(year, site, plot, species, willid, height) %>% 
  rename(yr = year, wilid = willid, pl_ht_cm = height) %>% 
  mutate(season = "fall") %>% 
  filter(wilid == "158")

# check specific records Tom identified

ht.combined %>%
  distinct() %>% 
  filter (yr == '2010' & site == "wb") %>%
  # datatable()
  filter(wilid == "158")

```

Two issues here, now fixed: 1.plant height of 1362 would appear to be a data entry error in the raw file. This is 2010 data drawn from the DCC (library repository), so I don't have the raw data to check. The value is in the raw file, not an artifact introduced in the data processing. I've changed it to "362".

```{r}

ht.combined %>%
  distinct() %>% 
  filter (yr == '2010' & site == "wb") %>%
  # datatable()
  filter(wilid == "158") %>% 
  datatable(caption = "questionable records flagged by TH")

```

1.Notice the different species: booth vs boothi, beb vs bebb. I ran code to find and replace such issues, but had it too early in the chain of transformations involved in the import and merging of the raw data. I have steps to remove duplicates, but b/c of the different species, they're not the same.

```{r}
## add full id to join with species
# wilid and site are not unique
ht.combined <- ht.combined %>% 
  filter(!is.na(pl_ht_cm))

```

```{r}
# ht.combined <- ht.combined %>% 
#   mutate(plot = case_when(site == "crescent" & plot == "cc" ~ "obs",
#                           site == "elk1" & plot == "cc" ~ "obs",
#                           site == "elk2" & plot == "cc" ~ "obs",
#                           site == "elk3" & plot == "cc" ~ "obs",
#                           site == "elk4" & plot == "cc" ~ "obs",
#                           site == "elk5" & plot == "cc" ~ "obs",
#                           site == "ghole" & plot == "cc" ~ "obs",
#                           site == "hailbuf" & plot == "cc" ~ "obs",
#                           site == "lava" & plot == "cc" ~ "obs",
#                           site == "lb" & plot == "cc" ~ "obs",
#                           site == "lb1" & plot == "cc" ~ "obs",
#                           site == "lb2" & plot == "cc" ~ "obs",
#                           site == "lb3" & plot == "cc" ~ "obs",
#                           site == "lb4" & plot == "cc" ~ "obs",
#                           site == "lostc" & plot == "cc" ~ "obs",
#                           site == "lostl" & plot == "cc" ~ "obs",
#                           site == "oxbow" & plot == "cc" ~ "obs",
#                           site == "rose" & plot == "cc" ~ "obs",
#                           site == "slide" & plot == "cc" ~ "obs",
#                           site == "wb1" & plot == "cc" ~ "obs",
#                           site == "wb2" & plot == "cc" ~ "obs",
#                           site == "wb4" & plot == "cc" ~ "obs",
#                           site == "yancy" & plot == "cc" ~ "obs",
#                           TRUE ~ plot
#                           ))
#   
```

```{r}

ht.combined %>% ggTile_yr_season_site2()

## cleaning
ht.combined %>% 
  filter(season  == "spring") %>% 
  tabyl(site,plot) %>% 
  gt()

# Eliminate species codes, find distinct, then join in the species from LU

# manually correct missing sites in raw files and
# regenerate the full_id
ht.combined <- ht.combined %>%
  mutate(site = case_when(wilid == "415" & is.na(site) ~ "ghole",
                      wilid == "418" & is.na(site) ~ "ghole",
                      wilid == "449" & is.na(site) ~ "glen",
                      wilid == "452" & is.na(site) ~ "glen",
                      TRUE ~ site)) %>%
  mutate(site2 = paste0(site,"-", plot)) %>% 
  mutate(wilid_full = paste0(site2,"-", wilid))

## still issues with wilid decimal points
ht.combined <-  ht.combined %>%
  mutate(wilid = str_remove(.$wilid, ".0")) 

## !!  
ht.combined <- ht.combined %>% 
  mutate(wilid = as.numeric(wilid)) %>% 
  filter(!is.na(wilid)) %>% 
  filter(wilid != "0") %>%
  mutate(wilid_full = paste0(site,"-",plot,"-",wilid)) %>% 
  distinct()

ht.combined %>% 
  ggTile_yr_season_site2()

# problem records: plot
ht.combined <- ht.combined %>% 
  # mutate(plot  = case_when(site == "crescent" & plot == "cc" ~ "obs",
  #                          TRUE ~ plot)) %>%
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid))
  
# ht.combined %>% visdat::vis_dat()

```

```{r}
ht.combined %>% 
  ggTile_yr_season_site2() +
  labs(caption = "ht.combined")

```

```{r, eval=FALSE}
# ### Bind in the sprcag and prod derived ht data
ht.combined2 <- ht.combined %>% 
  mutate(wilid = as.character(wilid)) %>% 
  bind_rows(., ht.all.01.19) %>% 
  distinct() 

ht.combined2 %>% ggTile_yr_season_site2()

ht.combined2 <- ht.combined %>%
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid))

ht.combined %>% 
  ggTile_yr_season_site2()

ht.combined %>% 
  filter(plot != "obs") %>% 
  ggTile_yr_season_site2() +
  labs(title = "exp sites")

ht.combined %>% 
  filter(plot == "obs") %>% 
  ggTile_yr_season_site2() +
  labs(title = "obs sites")

```

```{r}
## more filtering. I'll leave it to TH to remove
# ht.combined <- ht.combined %>% 
#   filter(pl_ht_cm < 600)

# fix glen
ht.combined <- ht.combined %>% 
  mutate(plot = case_when(site == "glen" & plot == "cc" ~ "obs",
                          TRUE ~ plot))
```

```{r}
### QA QC checks on combined data
# ht.combined %>%
#   distinct(wilid) %>% 
#   datatable(caption = "distinct wilid")

# ht.combined %>%
#   distinct(wilid_full) %>% 
#   datatable(caption = "distinct wilid_full")

ht.combined %>%
  distinct(site2) %>% 
  datatable(caption = "distinct site2")

```

```{r}
ht.combined %>% 
  mutate(yr = as.character(yr)) %>%
  filter(plot != "obs") %>% 
  ggPoint_yr_site_season_plot()

```

```{r, eval=FALSE, fig.height=10}
gg.plty1 <- ht.combined %>% 
  ggplot(aes(x = site2, y = pl_ht_cm)) +
  geom_point()  +
  coord_flip()

plotly::ggplotly(gg.plty1)

```

```{r, eval=FALSE}
## create site2 beaver occ lu
### export site 2 lu
# ht.combined.f3 %>% 
#   distinct(site, plot, site2) %>% 
#   write_csv("./data/lu_site_site2_beav.csv")
## manually update with LM input
```

```{r}
# Beaver currently occupy crystal (2015?-2019) and wb4 (2017-2019).
# 
# Lb4, lb5, and lb6 (2016/2017-2018), along with elk4 and elk5 (not sure of dates but not currently occupied). 

# lu_site_beav <- read_csv("./data/lu_site_site2_beav.csv")

```

```{r}
## plt join
ht.combined <- ht.combined %>%
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid))

ht.combined %>% 
  ggTile_yr_season_site2()
  
```

```{r}

ht.combined <- ht.combined %>% 
  filter(!(yr == "2006" & pl_ht_cm == 460)) %>%
  filter(pl_ht_cm > 1)

p <- ht.combined  %>% 
  # mutate(pl_ht_cm = if_else(condition = pl_ht_cm < 1, true = pl_ht_cm*100, false = pl_ht_cm)) %>% 
  filter(site == "eb1") %>% 
  ggplot(aes(yr,pl_ht_cm)) +
  geom_jitter(aes(color = plot)) +
  facet_wrap(~site2) 

plotly::ggplotly(p)

# pltly1  
pltly1 <- ht.combined %>% 
  filter(plot != "obs") %>% 
  ggplot(aes(yr,pl_ht_cm)) +
  geom_jitter(aes(color = plot)) +
  facet_wrap(~site2)

plotly::ggplotly(pltly1)

```

```{r}
# fix
ht.combined <- ht.combined %>%
  filter(!is.na(site))

ht.combined %>%
  filter(!is.na(site)) %>% 
  filter(wilid == "353") %>% 
  ggplot(aes(yr, pl_ht_cm)) +
  geom_point(aes(color = season)) +
  # geom_label_repel(aes(label = wilid)) +
  geom_label(aes(label = season), size = 2.5) +
   geom_point(aes(color = season), size = 6, alpha = .1) +
  facet_wrap(~site, ncol = 1) +
  theme_minimal() +
  # facet_grid(season~site) +
  labs(title = "wilid 353: suspect 'site' attribution for elk and elk1")

# make educated guess and reclass site for 2014
ht.combined <- ht.combined %>% 
  mutate(site = case_when(yr == "2014" & site == "elk" ~ "elk2",
                          TRUE ~ site))


ht.combined %>%
  filter(!is.na(site)) %>% 
  filter(wilid == "353") %>% 
  ggplot(aes(yr, pl_ht_cm)) +
  geom_point(aes(color = season)) +
  # geom_label_repel(aes(label = wilid)) +
  geom_label(aes(label = season), size = 2.5) +
   geom_point(aes(color = season), size = 6, alpha = .1) +
  facet_wrap(~site, ncol = 1) +
  # facet_grid(season~site) +
  labs(title = "wilid 353: suspect 'site' attribution for elk and elk1")

ht.combined %>%
  filter(!is.na(site)) %>% 
  filter(wilid == "353") %>% 
  ggplot(aes(yr, pl_ht_cm)) +
  geom_point(aes(color = season)) +
  # geom_label_repel(aes(label = wilid)) +
  geom_label(aes(label = wilid)) +
  facet_wrap(~site, ncol = 1) +
  # facet_grid(season~site) +
  labs(title = "wilid 353: suspect 'site' attribution for elk and elk1")

```

```{r}
# clean
# address issue of crystal and elk 2 as != obs
ht.combined <- ht.combined %>%
  mutate(plot = case_when(site == "crystal" ~ "obs",
                          TRUE ~ plot)) %>% 
  mutate(plot = case_when(site == "elk2" ~ "obs",
                          TRUE ~ plot)) %>% 
  mutate(site2 = paste0(site,"-",plot)) %>% 
  mutate(wilid_full = paste0(site2,"-",wilid)) %>% 
  # select(-c(species, spp_notes)) %>% 
  distinct()

```

```{r}

# gt.1 <- function(df){
#   df %>% 
#     gt::gt() %>% 
#     tab_header(title = "Count")
# }
# 
# ht.combined %>%
#   filter(plot != "obs") %>%
#   filter(season == "spring") %>% 
#   tabyl(site, plot, yr) %>%
#   # set_names() %>% 
#   map_df(.,rbind, .id = 'names') %>% 
#   gt() %>% 
#   tab_header(title = "Spring data - count of observations")

ht.combined %>%
  # filter(site == "elk2") %>% 
  filter(plot != "obs") %>%
  filter(season == "spring") %>% 
  tabyl(site, plot, yr) %>%
  # set_names() %>% 
  map_df(.,rbind, .id = 'yr') %>% 
  datatable(caption = "Spring data - count of observations", rownames = FALSE, filter = "top")

```

```{r, eval = FALSE}

ht.combined.nest <- ht.combined %>%
  # filter(wilid == "353")
  group_by(site, yr) %>% 
  nest()

ht.combined.nest %>%
  pluck(3) %>% 
  pluck(5) %>%
  gt()

```

```{r}
#### lu merge
# dcc.wilid.spp.lu
# wilid.spp.lu <- wilid.spp.lu %>% 
#   mutate(wilid_full = paste0(site, "-", plot, "-", wilid)) 

## create a lu with dcc raw and v1 wilid spp list
# wilid.spp.lu.2 <- bind_rows(wilid.spp.lu, dcc.wilid.spp.lu) %>% 
#   distinct()

# write_csv(wilid.spp.lu.2, "./data/wilid_spp._lu_v2.csv")

```

```{r, eval = FALSE, echo=FALSE}
# create updated lu
# ht.combined %>% 
#   distinct(wilid, site, treat, exp, dam, browse, site2) %>%
#   write_csv("./data/wilid_lookup_master_v2.csv")

## create species lookup
# nsf.ht.spr %>% 
#   distinct(site, plot, species, wilid) %>% 
#   write_csv("./data/wilid_spp_lu.csv")

```

## Plots

```{r,fig.height=10, fig.width=9, eval = FALSE, echo=FALSE}
### boxplots of Fall height measurements
# ht.combined %>% 
#   filter(site == "elk2" & yr == "2014")

ht.combined %>% 
  filter(season=="fall") %>% 
  filter(plot != "obs") %>% 
  ggplot(aes(x=yr,y=pl_ht_cm)) +
  geom_boxplot(aes(fill = plot)) +
  theme_minimal() +
  # theme(legend.position = "none") +
  labs(title = "Fall height measurements", y = "Height (cm)") +
  facet_wrap(~site, ncol=1, scales = "free") +
  theme(axis.text.x = element_text(face="bold", color="black", 
                           size=9, angle=75, hjust = 1))

ht.combined %>% 
  filter(season=="spring") %>% 
  filter(plot != "obs") %>% 
  ggplot(aes(x=yr,y=pl_ht_cm)) +
  geom_boxplot(aes(fill = plot)) +
  theme_minimal() +
  # theme(legend.position = "none") +
  labs(title = "Spring height measurements", y = "Height (cm)") +
  facet_wrap(~site, ncol=1, scales = "free") +
  theme(axis.text.x = element_text(face="bold", color="black", 
                           size=9, angle=75, hjust = 1))

```

```{r}
## summarize by site2 and year using skimr
summary.by.site2 <- ht.combined %>% 
  # filter(wilid == "924") %>% 
  filter(!is.na(site)) %>% 
  mutate(year = as.factor(yr)) %>% 
  group_by(site, plot, site2,yr) %>%
  nest() %>% 
  mutate(skim2 = map(data,skim)) %>% 
  unnest(skim2) %>% 
  ungroup()

```

```{r, eval=TRUE}
## line plot
summary.by.site2 %>%
  mutate(yr = as.integer(yr)) %>% 
  filter(skim_variable == 'pl_ht_cm') %>%
  filter(plot != "obs") %>% 
  # filter(!(site2 == "elk-dx" & yr == 2019)) %>% 
  ggplot(aes(yr)) +
  geom_line(aes(y=numeric.p50, color = plot)) +
  geom_point(aes(y=numeric.p50, color = plot)) +
  geom_pointrange(aes(y = numeric.p50, ymin = numeric.p25, ymax = numeric.p75, color = plot), alpha = 0.31) +
  facet_wrap(~site) +
  theme_minimal() +
  labs(y="Median willow height (cm)", x = "Year", caption = "Willow height at experimental sites")
# ggsave("./output/median_willow_height_20191205_0924.png", width = 7, height = 5.5)


```

```{r}
##
summary.by.site2 %>%
  select(yr, plot, site2, skim_variable, contains("numeric")) %>% 
  filter(!(site2 == "elk-dx" & yr == 2019)) %>% 
  filter(skim_variable == "pl_ht_cm") %>% 
  filter(plot != "obs") %>%
  mutate(yr = as.integer(yr)) %>% 
  ggplot(aes(yr, site2)) +
  geom_tile(aes(fill = numeric.mean), color = 'gray70') +
  scale_fill_viridis() +
  theme_light() +
  labs(x = "Year", y = "", subtitle = "Mean willow height", fill = "Height (cm)", caption = getwd())
# ggsave("./output/median_willow_height_HM_20191205_0924.png", width = 7, height = 7)

```

```{r, eval=FALSE}
## create species lookup
# ht.combined %>% 
#   mutate(species = case_when(species == 'beb' ~ "bebb",
#                              species == 'drumm' ~ "drum",
#                              species == 'boothii' ~ "booth",
#                              species == 'boothi' ~ "booth",
#                              species == 'lassiandra' ~ "lass",
#                              species == 'plantifolia' ~ "plan",
#                              species == 'planifolia' ~ "plan",
#                              
#                              TRUE ~ species)) %>% 
#   # filter(is.na(species))
#   # filter(species == "na")
#   distinct(site2, wilid, species) %>% 
#   filter(!is.na(species)) %>%
#   filter(species != "na") %>%
#   mutate(wilid_full = paste0(site2,"-",wilid)) %>%
#   select(wilid_full, species) %>% 
#   write_csv("./data/lu_spp_site2.csv")

```

```{r}
ht.combined %>% 
  # filter(is.na(species)) %>% 
  group_by(site, wilid, yr) %>% 
  tally() %>% 
  datatable(filter = "top")
  
```

## Species and plot lu join

```{r}

lu.wilid.spp.v3 <- lu.wilid.spp.v3 %>% 
  select(species, wilid_full) %>% 
  mutate(species = case_when(species == "beb" ~ "bebb",
                             species == "boothii" ~ "booth",
                             species == "boothi" ~ "booth",
                             species == "drum" ~ "drumm", 
                             species == " NA" ~ "NA",
                             species == "NA " ~ "NA",
                             species == "na" ~ "NA",
                             species == "plantifolia" ~ "plan",
                             species == "planifolia" ~ "plan",
                             species == "?" ~ "NA",
                             TRUE ~ species
                             ))

lu.wilid.spp.v3 %>% 
  distinct(species) %>% 
  gt() %>% 
  tab_header(title = "different species attributes")



### lu.wilid.spp.v3
# write_csv(lu.wilid.spp.v3, "./data/lu_wilid_full_spp_v3.csv")


```

```{r}
### spp join on wilid ful
ht.combined <- ht.combined %>% 
  left_join(.,lu.wilid.spp.v3, by = "wilid_full")

ht.combined <- ht.combined %>% 
  left_join(.,plot.lu) %>% 
  select(-treat)

ht.combined %>% 
  visdat::vis_dat()

## join bv beav
ht.combined <- ht.combined %>% 
  left_join(., lu_site_beav) 

ht.combined %>% visdat::vis_dat()

ht.combined %>% 
  group_by(yr, site, plot) %>% 
  tally() %>% 
  ggplot(aes(yr,site)) +
  geom_tile(aes(fill = n)) +
  facet_wrap(~plot)

```

## Export

```{r}

ht.combined %>% 
  # mutate(wilid = str_remove(.$wilid, ".0"))
  filter(is.na(species))
```

```{r, eval=FALSE}
ht.combined %>%
  # select(-c(species, spp_notes)) %>% 
  write_csv("./output/processed_data/ht_combined20200107_1618.csv")

```

# TH: missing list

TH sent pdf of tallies of questionable records on 20201118

```{r}
library(tabulizer)
tom.tbl <- tabulizer::extract_tables(file = "Willow_id_inventory_TH_20201118.pdf")

## bind matrices corresponding to each page in pdf
tom.tbl <- do.call(rbind, tom.tbl[-length(tom.tbl)])

## convert to tibble
tom.tbl <- tom.tbl %>% 
  as.tibble() %>% 
  janitor::row_to_names(1) %>% 
  janitor::clean_names() %>% 
  rename(row = x)

## write this table to csv
# write_csv(tom.tbl, "Willow_id_inventory_TH_20201118.csv")


tom.tbl %>% 
  datatable(rownames = FALSE, caption = "Extract of TH's 'Willow_id_inventory.pdf' file sent 20201118")

```

```{r}
##
```


```{r}
ht.combined %>% 
  filter(season == "fall") %>% 
  group_by(wilid_full, yr) %>% 
  summarise(n = n()) %>%
  datatable()
  
ht.combined %>% 
  distinct(yr)

## table similar to tom's
fa.tally.qa <- ht.combined %>%
  mutate(yr = as.integer(yr)) %>% 
  filter(season == "fall") %>% 
  group_by(wilid_full, yr) %>% 
  summarise(n = n()) %>%
  group_by(wilid_full) %>% 
  mutate(min.yr = min(yr, na.rm = TRUE), max.yr = max(yr, na.rm = TRUE)) %>%
  ungroup()

# fa.tally.qa %>% 
#   datatable()

fa.tally.qa %>% 
  filter(min.yr > "2015") %>% 
  datatable()

fa.tally.qa %>% 
  filter(min.yr > "2015") %>%
  mutate(min.yr = as.character(min.yr)) %>% 
  ggplot(aes(reorder(wilid_full, min.yr),yr)) +
  geom_tile(aes(fill = min.yr))

```

# Notes

The metadata csv included in the DCC repository does not include all the variables.

In the DCC data, "browse = 0" indicates experimental exclosure treatment. For example, see willid 1 in EB1. The plot type is listed in the 2018 dataset as "dx". In the 2001-2017 data, it's dam == 1, browse == 0

In the supplemental material accompanying the 2013 Marshall et al. ProcRoySoc paper, the willid are attributed with "dammed" and "UNbrowsed". This is confusing, since for "dam", 1 is the experimental treatment, but for "browse", 1 is the control.

```{r, echo = FALSE, eval = FALSE}
## Derived data export
## Write export file to disk


```
